### 2025-02-24 #7841 @editorialbot
[PRE REVIEW]: ScrollStats: a Python tool for quantifying scroll bar morphology on meandering rivers
**Submitting author:** @a-vanderheiden (<a href="http://orcid.org/0009-0003-2466-8144">Andrew Vanderheiden</a>)
**Repository:** https://github.com/tamu-edu/scrollstats
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
