### 2025-03-22 #7841 @a-vanderheiden
Hi @kthyng, thank you for reaching out and changing the track of this submission.

The main issue I see from the editorialbot are with the DOIs I included. Editorial bot says that there are skipped and missing DOIs but they all work for me when I check them individually in the browser. Is there something I need to change in my [`paper.bib`](https://github.com/tamu-edu/scrollstats/blob/main/paper/paper.bib) file?

I filtered the reviewer database for reviewers tagged with python and geomorphology in the ESE track which resulted in the following 6 reviewers:
- laurenguido
- WrenRaming
- ldominguezruben
- katmratliff
- kellykochanski
- rmsare
