### 2025-04-02 #7841 @kthyng
> The main issue I see from the editorialbot are with the DOIs I included. Editorial bot says that there are skipped and missing DOIs but they all work for me when I check them individually in the browser. Is there something I need to change in my [`paper.bib`](https://github.com/tamu-edu/scrollstats/blob/main/paper/paper.bib) file?

Check out the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) — looks like you should put your doi's in as `doi` instead of `url` and also skip the URL, just instead the doi itself.
