### 2025-04-02 #7841 @kthyng
@a-vanderheiden Some comments:
* I'd love to see this package set up on PyPI and/or conda-forge. Is there anything blocking you from doing that?
* While your readme is nicely laid out, I'd like to see real docs set up, as in something like a readthedocs. From what you have you'd need to just set up the infrastructure for readthedocs but looks like you already have the content written. Using sphinx can take advantage of the docstrings you've written to lay out your API nicely too.

Here are some resources that can help with these requests (a lot of this you already have but there should be some relevant bits in here):
* https://learn.scientific-python.org/development/guides/
* https://www.pyopensci.org/python-package-guide/
* https://www.pyopensci.org/python-package-guide/package-structure-code/publish-python-package-pypi-conda.html#
