+++ paper/paper.bib
@@ -2,7 +2,7 @@
 @article{hickin:1974,
 	title = {The development of meanders in natural river-channels},
 	volume = {274},
-	url = {https://doi.org/10.2475/ajs.274.4.414},
+	doi = {10.2475/ajs.274.4.414},
 	pages = {414--442},
 	number = {4},
 	journaltitle = {American Journal of Science},
@@ -14,7 +14,7 @@
 @article{nagy:2020,
 	title = {Point-bar Development Under Human Impact: Case Study on the {Lower Tisza River, Hungary}},
 	volume = {24},
-	url = {https://doi.org/10.5937/gp24-23011},
+	doi = {10.5937/gp24-23011},
 	pages = {1--12},
 	number = {1},
 	journaltitle = {Geographica Pannonica},
@@ -25,7 +25,7 @@
 @article{strick:2018,
 	title = {Morphology and spacing of river meander scrolls},
 	volume = {310},
-	url = {https://doi.org/10.1016/j.geomorph.2018.03.005},
+	doi = {10.1016/j.geomorph.2018.03.005},
 	pages = {57--68},
 	journaltitle = {Geomorphology},
 	author = {Strick, Robert J. P. and Ashworth, Philip J. and Awcock, Graeme and Lewin, John},
@@ -35,7 +35,7 @@
 @article{zen:2017,
 	title = {Exploring the role of trees in the evolution of meander bends: The {Tagliamento River, Italy}},
 	volume = {53},
-	url = {https://doi.org/10.1002/2017WR020561},
+	doi = {10.1002/2017WR020561},
 	pages = {5943--5962},
 	number = {7},
 	journaltitle = {Water Resources Research},
