### 2025-04-04 #7841 @a-vanderheiden
@kthyng

> - I'd love to see this package set up on PyPI and/or conda-forge. Is there anything blocking you from doing that?

I haven't encountered a specific block for this, I just haven't looked into it yet. I'll look into this and see what I can do!

> - While your readme is nicely laid out, I'd like to see real docs set up, as in something like a readthedocs. From what you have you'd need to just set up the infrastructure for readthedocs but looks like you already have the content written. Using sphinx can take advantage of the docstrings you've written to lay out your API nicely too.

Likewise, I haven't considered doing something like this before, but I can look into setting this up.

---
Thank you for the links! They were really helpful in wrapping my head around my next steps to get scrollstats properly packaged and documented.

In general, it looks like my steps are:
1. create a `pyproject.toml` file following the [simple package](https://learn.scientific-python.org/development/guides/packaging-simple/) method
2. Follow the [instructions from PyPI](https://packaging.python.org/en/latest/tutorials/packaging-projects/) to make sure that the rest of scrollstats is in the proper structure
3. [use sphinx create the docs ](https://www.sphinx-doc.org/en/master/tutorial/getting-started.html)
4. [generate the distribution archives](https://packaging.python.org/en/latest/tutorials/packaging-projects/#generating-distribution-archives) and upload to PyPI
5. [import scrollstats to readthedocs](https://docs.readthedocs.com/platform/stable/tutorial/#importing-the-project-to-read-the-docs)

Correct?
