### 2025-04-04 #7841 @kthyng
@a-vanderheiden Your list looks like a reasonable way to go. At the risk of muddying things, you might consider now or in a future project starting from a cookie cutter template, for example https://github.com/scientific-python/cookie from https://learn.scientific-python.org/development/. This sets you up with all the files you need to more easily get up and running with docs and packaging details without putting in as much work. The reason I bring this up is that once you're set up with GitHub Actions to, say, release to PyPI every time you make a new release, you don't need to do any of those manual steps by hand because they're built into the GitHub Action. You could also just set up the github action files yourself without using the cookie cutter template.
