### 2025-04-09 #7841 @a-vanderheiden
@kthyng Thank you for the recommendations! I'll get started with the scientific-python cookie cutter template and go from there. Getting these release processes automated with GH actions would be nice.
