### 2025-04-18 eb24b78
added readme badges from sp cookiecutter
