+++ README.md
@@ -1,5 +1,31 @@
 # ScrollStats
 
+[![Actions Status][actions-badge]][actions-link]
+[![Documentation Status][rtd-badge]][rtd-link]
+
+[![PyPI version][pypi-version]][pypi-link]
+[![Conda-Forge][conda-badge]][conda-link]
+[![PyPI platforms][pypi-platforms]][pypi-link]
+
+[![GitHub Discussion][github-discussions-badge]][github-discussions-link]
+
+<!-- SPHINX-START -->
+
+<!-- prettier-ignore-start -->
+[actions-badge]:            https://github.com/tamu-edu/scrollstats/workflows/CI/badge.svg
+[actions-link]:             https://github.com/tamu-edu/scrollstats/actions
+[conda-badge]:              https://img.shields.io/conda/vn/conda-forge/scrollstats
+[conda-link]:               https://github.com/conda-forge/scrollstats-feedstock
+[github-discussions-badge]: https://img.shields.io/static/v1?label=Discussions&message=Ask&color=blue&logo=github
+[github-discussions-link]:  https://github.com/tamu-edu/scrollstats/discussions
+[pypi-link]:                https://pypi.org/project/scrollstats/
+[pypi-platforms]:           https://img.shields.io/pypi/pyversions/scrollstats
+[pypi-version]:             https://img.shields.io/pypi/v/scrollstats
+[rtd-badge]:                https://readthedocs.org/projects/scrollstats/badge/?version=latest
+[rtd-link]:                 https://scrollstats.readthedocs.io/en/latest/?badge=latest
+
+<!-- prettier-ignore-end -->
+
 An open-source python library to calculate and extract morphometrics from scrollbar floodplains.
 
 
