+++ src/scrollstats/__init__.py
@@ -1,0 +1,11 @@
+"""
+Copyright (c) 2025 Andrew Vanderheiden. All rights reserved.
+
+scrollstats: An open-source python library to calculate and extract morphometrics from scroll bar floodplains
+"""
+
+from __future__ import annotations
+
+from ._version import version as __version__
+
+__all__ = ["__version__"]
