### 2025-04-19 de3b7c6
added nox to env
