### 2025-04-19 e5379d3
added core dependencies from environment.yml (conda)
