### 2025-04-21 b805312
add nbsphinx as docs dependency
