+++ pyproject.toml
@@ -53,6 +53,7 @@ docs = [
   "sphinx_copybutton",
   "sphinx_autodoc_typehints",
   "furo>=2023.08.17",
+  "nbsphinx",
 ]
 
 [project.urls]
