### 2025-04-21 c54807d
added config lines for api documentation according to sp docs
