### 2025-04-21 90142f9
added nbsphinx configurations to conf.py
