### 2025-04-22 #7841 @kthyng
@a-vanderheiden Sorry for my delay, I was out for a long weekend. I am not sure about the copyright question — that is more of a general question than something we dig into specifically. My guess is you would use 2023–2025 for the copyright.

I don't mean to rush you on working on this, but with a time guess I can manage things from my side.
