### 2025-04-22 #7841 @elbeejay
@kthyng in this case I might have a conflict of interest as well with one of the co-authors. Might be simpler to have another editor handle this - regardless of the timeline.
