+++ paper/paper.md
@@ -1,6 +1,8 @@
 ---
-title: 'ScrollStats: a Python tool for quantifying scroll bar morphology on meandering rivers'
-tags: 
+title:
+  "ScrollStats: a Python tool for quantifying scroll bar morphology on
+  meandering rivers"
+tags:
   - Python
   - scroll bars
   - river
@@ -8,34 +10,77 @@ tags:
   - geomorphology
 authors:
   - name: Andrew Vanderheiden
-    orcid: 0009-0003-2466-8144 
+    orcid: 0009-0003-2466-8144
     corresponding: true
-    affiliation: '1'
+    affiliation: "1"
   - name: Billy Hales
-    orcid: 0000-0002-6591-9363 
-    affiliation: '1'
+    orcid: 0000-0002-6591-9363
+    affiliation: "1"
   - name: Andrew J. Moodie
     orcid: 0000-0002-6745-036X
-    affiliation: '1'
-  - name: İnci Güneralp 
+    affiliation: "1"
+  - name: İnci Güneralp
     orcid: 0000-0002-7268-2974
-    affiliation: '1'
+    affiliation: "1"
 affiliations:
-- name: Department of Geography, Texas A&M University, United States
-  index: '1'
+  - name: Department of Geography, Texas A&M University, United States
+    index: "1"
 date: 14 February 2025
 bibliography: paper.bib
-
 ---
+
 # Summary
-Scroll bars are elongated arcuate topographic features deposited along the inner bank of meandering rivers. As the river continues to meander across the floodplain, the series of scroll bars deposited in its wake is known as ridge and swale topography. The ridge and swale topography, readily observed in LiDAR-derived digital elevation models (DEMs), contains a visually intuitive record of the river's migration history, and the specific morphology of each ridge may serve as a proxy for the hydrological, geomorphological, and sedimentological conditions under which each individual scroll bar is formed. The ridge crests, with their higher elevation relative to the swales, also encourage the growth of colonizing vegetation, which in turn stabilizes the ridges and mitigate future erosion [@zen:2017]. While there has long been an interest in the formation and preservation of scroll bars, research into the specific drivers of ridge morphology and what information it may contain is new [@strick:2018; @nagy:2020]. 
+
+Scroll bars are elongated arcuate topographic features deposited along the inner
+bank of meandering rivers. As the river continues to meander across the
+floodplain, the series of scroll bars deposited in its wake is known as ridge
+and swale topography. The ridge and swale topography, readily observed in
+LiDAR-derived digital elevation models (DEMs), contains a visually intuitive
+record of the river's migration history, and the specific morphology of each
+ridge may serve as a proxy for the hydrological, geomorphological, and
+sedimentological conditions under which each individual scroll bar is formed.
+The ridge crests, with their higher elevation relative to the swales, also
+encourage the growth of colonizing vegetation, which in turn stabilizes the
+ridges and mitigate future erosion [@zen:2017]. While there has long been an
+interest in the formation and preservation of scroll bars, research into the
+specific drivers of ridge morphology and what information it may contain is new
+[@strick:2018; @nagy:2020].
 
 # Statement of need
-`ScrollStats` is an open-source Python tool to quantify the morphology of scroll bars preserved in the ridge and swale topography commonly found in the floodplains of meandering rivers adjacent to the river channel. This quantification will allow researchers to investigate the relationships between ridge morphology and the environmental factors affecting its formation, such as the hydrology at the time of deposition, spatial variations in the river width, the channel curvature, the position along the meander bend, and the floodplain vegetation coverage and composition.
 
[…]
