+++ src/scrollstats/delineation/ridge_area_raster.py
@@ -1,42 +1,42 @@
-from typing import Callable, Tuple, Dict
-from functools import partial
-from pathlib import Path
-from inspect import signature
+from __future__ import annotations
 
+from functools import partial
+from inspect import signature
+from pathlib import Path
+from typing import Callable, Dict, Tuple
+
+import geopandas as gpd
 import numpy as np
 import rasterio
 import rasterio.mask
-import geopandas as gpd
 from shapely.geometry import Polygon
 
 from .array_types import (
-    Array2D, 
-    ElevationArray2D, 
+    Array2D,
     BinaryArray2D,
     BinaryClassifierFn,
-    BinaryDenoiserFn
+    BinaryDenoiserFn,
+    ElevationArray2D,
 )
-
 from .raster_classifiers import DEFAULT_CLASSIFIERS
 from .raster_denoisers import DEFAULT_DENOISERS
 
 
-def clip_raster(ds:rasterio.DatasetReader, geometry:Polygon, array=None, no_data=None) -> Tuple[ElevationArray2D, BinaryArray2D, Dict]:
-
+def clip_raster(
+    ds: rasterio.DatasetReader, geometry: Polygon, array=None, no_data=None
+) -> Tuple[ElevationArray2D, BinaryArray2D, Dict]:
     # Replace optional values
     if isinstance(array, np.ndarray):
         array_copy = array.copy()
     else:
         array_copy = ds.read(1)
-    
+
     if not no_data:
         no_data = ds.nodata
 
     # For cropped_mask, True is area outside of geometry
-    clipped_mask, transform, window= rasterio.mask.raster_geometry_mask(
-        dataset=ds,
-        shapes=[geometry],
-        crop=True
+    clipped_mask, transform, window = rasterio.mask.raster_geometry_mask(
+        dataset=ds, shapes=[geometry], crop=True
     )
 
     # Update size, transform, and nodata value for output raster
@@ -53,7 +53,7 @@ def clip_raster(ds:rasterio.DatasetReader, geometry:Polygon, array=None, no_data
 
     # Crop array
     array_clip = array_copy[window.toslices()]
-    
+
     # Fill no_data values for output array
     # Only 0s are falsy, so any other number (including np.nan) will evaluate to True in boolean arrays.
     # So, we need to explicitly set the fill value as False for boolean arrays.
@@ -66,53 +66,55 @@ def clip_raster(ds:rasterio.DatasetReader, geometry:Polygon, array=None, no_data
     return array_clip, clipped_mask, clipped_meta
 
 
-def partial_from_kwargs(func:Callable, **kwargs) -> Callable:
-    """Create a partial function from all of the kwargs that are found in the function's signature """
+def partial_from_kwargs(func: Callable, **kwargs) -> Callable:
+    """Create a partial function from all of the kwargs that are found in the function's signature"""
     sig = signature(func)
-    args = {p:kwargs.get(p) for p in sig.parameters if kwargs.get(p)}
+    args = {p: kwargs.get(p) for p in sig.parameters if kwargs.get(p)}
     return partial(func, **args)
 
 
-def create_ridge_area_raster(dem_ds:rasterio.DatasetReader, 
-                             geometry:Polygon, 
-                             classifier_funcs:Tuple[BinaryClassifierFn] = DEFAULT_CLASSIFIERS,
-                             denoiser_funcs:Tuple[BinaryDenoiserFn] = DEFAULT_DENOISERS,
-                             no_data_value = None,
-                             **kwargs) -> tuple[Array2D, ElevationArray2D, Dict]:
+def create_ridge_area_raster(
[…]
