+++ src/scrollstats/delineation/array_types.py
@@ -1,5 +1,7 @@
+from __future__ import annotations
 
-from typing import Tuple, Callable
+from typing import Callable, Tuple
+
 import numpy as np
 
 # Define types
