### 2025-04-25 002484b
snake_case rename for ridge_amplitude
