### 2025-04-25 b4df9a9
pylint errors: raise Exception
