### 2025-04-26 e88bc93
explicitly iterate on dict.keys() + other pylint fixes
