### 2025-04-26 f942dc8
docstring update + unused variable notation
