### 2025-04-26 e92efec
removed unnecessary use of a comprehension
