### 2025-04-26 4cc4783
fixed: variable name "lu_A" doesn't conform to snake_case naming style
