+++ src/scrollstats/delineation/raster_classifiers.py
@@ -111,7 +111,7 @@ def parameter_quadratic_planform_curvature(_coeff):
     -------
     Single parameter for local planform curvature.
     """
-    a, b, c, d, e, f = _unpack_coeff(_coeff)
+    a, b, c, d, e, _f = _unpack_coeff(_coeff)
 
     if (d == 0) and (e == 0):
         _curv_param = 0.0
