### 2025-04-26 3cdda95
fixed: Variable name "N" doesn't conform to snake_case naming style
