+++ src/scrollstats/ridge_metrics/ridge_amplitude.py
@@ -20,7 +20,7 @@ def calc_ridge_maxes(dem_sig, mask):
 
 
 def calc_swale_mins(dem_sig, mask):
-    """Calcualte the minimun value within each of the swale areas of the mask"""
+    """Calculate the minimum value within each of the swale areas of the mask"""
 
     # Find each unique swale
     labels, numfeats = ndimage.label(~mask)
@@ -34,18 +34,18 @@ def calc_swale_mins(dem_sig, mask):
 
 def s00_amps(maxes, mins):
     """
-    Amplitude calcualtion if binary signal is capped with 0s on each end.
+    Amplitude calculation if binary signal is capped with 0s on each end.
           _   _   _
     ex. _| |_| |_| |_
 
-    Amplitude for each ridge is calcualted as the average of the differences between ridge max and the swale mins on each side.
+    Amplitude for each ridge is calculated as the average of the differences between ridge max and the swale mins on each side.
     However, the amplitude of the ridge at the beginning is calculated only by the following swale because of channel effects on the DEM signal.
     """
 
-    # Calcualte diffs between ridge maxes and the preceeding swale mins
+    # Calculate diffs between ridge maxes and the preceding swale mins
     d1 = maxes[1:] - mins[1:-1]
 
-    # Calcualte diffs between ridge maxes and following swale mins
+    # Calculate diffs between ridge maxes and following swale mins
     d2 = maxes - mins[1:]
 
     # Insert the first value from d2 into d1 to make arrays the same size
@@ -57,19 +57,19 @@ def s00_amps(maxes, mins):
 
 def s11_amps(maxes, mins):
     """
-    Amplitude calcualtion if binary signal is capped with 1s on each end.
+    Amplitude calculation if binary signal is capped with 1s on each end.
         _   _   _
     ex.  |_| |_|
 
-    Amplitude for each ridge is calcualted as the average of the differences between ridge max and the swale mins on each side.
+    Amplitude for each ridge is calculated as the average of the differences between ridge max and the swale mins on each side.
     However, because there is a ridge at both ends, the first and last ridge only gets one amplitude measurement
 
     """
 
-    # Calcualte diffs between ridge maxes and the preceeding swale mins
+    # Calculate diffs between ridge maxes and the preceding swale mins
     d1 = maxes[1:] - mins
 
-    # Calcualte diffs between ridge maxes and following swale mins
+    # Calculate diffs between ridge maxes and following swale mins
     d2 = maxes[:-1] - mins
 
     # Insert the first value from d2 because d1 is missing a measurement for the first ridge max
@@ -84,19 +84,19 @@ def s11_amps(maxes, mins):
 
 def s01_amps(maxes, mins):
     """
-    Amplitude calcualtion if binary signal starts with a swale and ends with a ridge.
+    Amplitude calculation if binary signal starts with a swale and ends with a ridge.
           _   _   _
     ex. _| |_| |_|
 
-    Amplitude for each ridge is calcualted as the average of the differences between ridge max and the swale mins on each side.
+    Amplitude for each ridge is calculated as the average of the differences between ridge max and the swale mins on each side.
     However, because there is no swale at the end, the last ridge only gets one amplitude measurement
     Additionally, the amplitude of the ridge at the beginning is calculated only by the following swale because of channel effects on the DEM signal.
     """
 
-    # Calcualte diffs between ridge maxes and the preceeding swale mins
+    # Calculate diffs between ridge maxes and the preceding swale mins
     d1 = maxes[1:] - mins[1:]
 
-    # Calcualte diffs between ridge maxes and following swale mins
+    # Calculate diffs between ridge maxes and following swale mins
     d2 = maxes[:-1] - mins[1:]
 
[…]
