+++ .pre-commit-config.yaml
@@ -61,6 +61,8 @@ repos:
     rev: "v2.4.1"
     hooks:
       - id: codespell
+        additional_dependencies:
+          - tomli
 
   - repo: https://github.com/shellcheck-py/shellcheck-py
     rev: "v0.10.0.1"
