+++ pyproject.toml
@@ -168,3 +168,6 @@ messages_control.disable = [
   "missing-function-docstring",
   "wrong-import-position",
 ]
+
+[tool.codespell]
+ignore-words-list = "FLUD" # lab acronym 
