### 2025-04-26 eeea9ea
codespell exceptions
