### 2025-04-27 77ba7ba
iterate over .items() explicitly
