### 2025-04-27 d097edd
ignore mypy errors from untyped 3rd party libraries
