### 2025-04-27 0ae7b80
set explicit module exports
