+++ src/scrollstats/delineation/__init__.py
@@ -1,5 +1,13 @@
 from __future__ import annotations
 
+__all__ = [
+    "LineSmoother",
+    "create_ridge_area_raster",
+    "create_ridge_area_raster_fs",
+    "quadratic_profile_curvature",
+    "residual_topography",
+]
+
 from .line_smoother import LineSmoother
 from .raster_classifiers import quadratic_profile_curvature, residual_topography
 from .ridge_area_raster import create_ridge_area_raster, create_ridge_area_raster_fs
