+++ src/scrollstats/transecting/__init__.py
@@ -1,3 +1,8 @@
 from __future__ import annotations
 
+__all__ = [
+    "MultiTransect",
+    "create_transects",
+]
+
 from .transect import MultiTransect, create_transects
