+++ src/scrollstats/ridge_metrics/__init__.py
@@ -1,5 +1,14 @@
 from __future__ import annotations
 
+__all__ = [
+    "BendDataExtractor",
+    "RidgeDataExtractor",
+    "TransectDataExtractor",
+    "calc_ridge_amps",
+    "calculate_ridge_metrics",
+    "map_amp_values",
+]
+
 from .calc_ridge_metrics import calculate_ridge_metrics
 from .data_extractors import (
     BendDataExtractor,
