### 2025-04-29 39e5c77
type annotations for calc_ridge_metircs()
