### 2025-04-29 3d03884
fixed calls to untyped functions in test_core.py
