### 2025-04-29 d22d6bf
added parameters for generics and added missing type annotations
