### 2025-04-29 a4dc30a
added logic to handle potential 2nd return array
