### 2025-04-29 161b549
added function return types
