### 2025-04-29 76c8949
added type annotations for function parameters
