### 2025-04-30 5c3a1d8
added type annotations for parameters and return values
