+++ src/scrollstats/ridge_metrics/data_extractors.py
@@ -11,9 +11,13 @@ Rules:
 
 from __future__ import annotations
 
+from collections.abc import Iterable
+from typing import Any
+
 import geopandas as gpd
 import numpy as np
 import pandas as pd
+from rasterio.io import DatasetReader as RasterioDatasetReader
 from scipy import ndimage
 from shapely.geometry import LineString, Point
 from tqdm import tqdm
@@ -21,12 +25,12 @@ from tqdm import tqdm
 from .ridge_amplitude import calc_ridge_amps
 
 
-def calc_dist(p1: np.array, p2: np.array) -> np.array:
+def calc_dist(p1: np.ndarray, p2: np.ndarray) -> np.ndarray:
     """p1 and p2 are both (n,2) arrays of coordinates"""
     return np.sqrt((p1[:, 0] - p2[:, 0]) ** 2 + (p1[:, 1] - p2[:, 1]) ** 2)
 
 
-def remove_coincident_points(coord_array: np.array) -> np.array:
+def remove_coincident_points(coord_array: np.ndarray) -> np.ndarray:
     """Removes coincident points that happen to be adjacent"""
 
     # Calc distances between each point
@@ -40,12 +44,12 @@ def remove_coincident_points(coord_array: np.array) -> np.array:
     return coord_array[unique_idx]
 
 
-def explode(line):
+def explode(line: LineString) -> list[LineString]:
     """Return a list of all 2 coordinate line segments in a given LineString"""
     return list(map(LineString, zip(line.coords[:-1], line.coords[1:])))
 
 
-def densify_segment(line):
+def densify_segment(line: LineString) -> LineString:
     """Densify a line segment between two points to 1pt/m. Assumes line coordinates are in meters"""
     x, y = line.xy
 
@@ -57,7 +61,7 @@ def densify_segment(line):
     return LineString(zip(xs, ys))
 
 
-def densify_line(line):
+def densify_line(line: LineString) -> LineString:
     """Return the given LineString densified coordinates. Density = 1pt/unit length"""
 
     all_points = []
@@ -77,7 +81,9 @@ def densify_line(line):
     return LineString(unique_points)
 
 
-def transform_coords(coord_array, bin_raster):
+def transform_coords(
+    coord_array: np.ndarray, bin_raster: RasterioDatasetReader
+) -> np.ndarray:
     """Transform the coordinates from geo to image for indexing a in_bin_raster"""
 
     transform = bin_raster.transform
@@ -93,7 +99,12 @@ class RidgeDataExtractor:
     """
 
     def __init__(
-        self, geometry, position, ridges, dem_signal=None, bin_signal=None
+        self,
+        geometry: LineString,
+        position: int,
+        ridges: gpd.GeoDataFrame,
+        dem_signal: np.ndarray[float] | None = None,
+        bin_signal: np.ndarray[float] | None = None,
     ) -> None:
         # Inputs
         self.id = None
@@ -143,13 +154,13 @@ class RidgeDataExtractor:
 
         self.gdf = self.coerce_dtypes(self.gdf)
 
-    def determine_signal_length(self):
+    def determine_signal_length(self) -> float:
         """Return length of dem/bin signal if provided"""
         if isinstance(self.bin_signal, np.ndarray):
             return self.dem_signal.size
         return np.nan
 
-    def create_point_gdf(self):
+    def create_point_gdf(self) -> gpd.GeoDataFrame:
         """Create a 3 point GeoDataFrame to contain all relevant info for other methods."""
         gdf = gpd.GeoDataFrame(
             columns=self.data_columns, geometry="geometry", crs=self.ridges.crs
@@ -157,7 +168,7 @@ class RidgeDataExtractor:
         gdf = gdf.set_index("p_id")
         return gdf
 
-    def add_point_geometries(self, gdf, line):
[…]
