### 2025-05-01 8e411af
added type annotations for parameters and return values
