### 2025-05-01 3853493
added type annotations for parameters and return values
