### 2025-05-03 09f6a25
added type annotations for parameters and return values
