+++ src/scrollstats/transecting/transect.py
@@ -10,12 +10,13 @@ from shapely.geometry import (
     MultiPoint,
     MultiPolygon,
     Point,
+    Polygon,
 )
 from tqdm import tqdm
 
 
 ## Simple Geometric functions: accepts and returns arrays of coordinates
-def curvature(line: LineString) -> Tuple[np.ndarray, np.ndarray]:
+def curvature(line: LineString) -> Tuple[np.ndarray[float], np.ndarray[float]]:
     x, y = line.xy
 
     # Get dx, dy
@@ -48,7 +49,7 @@ def curvature(line: LineString) -> Tuple[np.ndarray, np.ndarray]:
     return curv, a
 
 
-def calc_dist_along_line(line: LineString) -> np.array:
+def calc_dist_along_line(line: LineString) -> np.ndarray[float]:
     """Calculate an array of distance along a given line"""
 
     x, y = line.xy
@@ -61,12 +62,14 @@ def calc_dist_along_line(line: LineString) -> np.array:
     return np.cumsum(arc_length)
 
 
-def calc_dist(p1: np.array, p2: np.array) -> np.array:
+def calc_dist(p1: np.ndarray[float], p2: np.ndarray[float]) -> np.ndarray[float]:
     # p1 and p2 are both (n,2) arrays of coordinates
     return np.sqrt((p1[:, 0] - p2[:, 0]) ** 2 + (p1[:, 1] - p2[:, 1]) ** 2)
 
 
-def law_of_cos(a: np.array, b: np.array, c: np.array) -> np.array:
+def law_of_cos(
+    a: np.ndarray[float], b: np.ndarray[float], c: np.ndarray[float]
+) -> np.ndarray[float]:
     """
     This function uses law of cosines to calc the interior angle formed by two connected lines
     Traveling down the line, vertices are a, b, c (a=beg, b=mid, c=end)
@@ -84,7 +87,7 @@ def law_of_cos(a: np.array, b: np.array, c: np.array) -> np.array:
     return beta
 
 
-def calc_dxdy(angle: np.array, length=100) -> np.array:
+def calc_dxdy(angle: np.ndarray[float], length: float = 100) -> np.ndarray[float]:
     """Calculates the delta x and y given an angle and a length"""
 
     dx = np.cos(angle) * length
@@ -93,7 +96,9 @@ def calc_dxdy(angle: np.array, length=100) -> np.array:
     return np.vstack([dx, dy])
 
 
-def add_sub_90(a: Point, b: Point, alpha: float, length=100) -> np.array:
+def add_sub_90(
+    a: Point, b: Point, alpha: float, length: float = 100
+) -> np.ndarray[float]:
     """
     This function determines which direction to shoot a new transect (+ or - 90°)
     The proper line should have the largest of the interior angles formed by the incoming transect and new shoot being formed
@@ -131,7 +136,7 @@ def add_sub_90(a: Point, b: Point, alpha: float, length=100) -> np.array:
     return coords[max_idx]
 
 
-def vert_res(p0: Point, p1: Point, p2: Point) -> np.array:
+def vert_res(p0: Point, p1: Point, p2: Point) -> np.ndarray[float]:
     """Calculates the vertical resultant of the two vectors (LineStrings) p0->p1 and p0->p2"""
     return np.asarray(p1.xy) + np.asarray(p2.xy) - np.asarray(p0.xy)
 
@@ -149,7 +154,7 @@ def find_closest_idx(point: Point, line: LineString) -> int:
     return s_tups[0][0]
 
 
-def direction_alpha_at_point(point: Point, line: LineString) -> tuple:
+def direction_alpha_at_point(point: Point, line: LineString) -> tuple[int, float]:
[…]
