### 2025-05-04 e013b5c
added nptyping to fix array types
