+++ src/scrollstats/delineation/array_types.py
@@ -1,17 +1,17 @@
 from __future__ import annotations
 
-from typing import Callable, Tuple
+from typing import Any, Callable
 
-import numpy as np
+from nptyping import NDArray, Shape
 
-# Define types
-Array2D = np.ndarray[Tuple[int, int], np.dtype[any]]
-ElevationArray2D = np.ndarray[Tuple[int, int], np.dtype[float]]
-BinaryArray2D = np.ndarray[Tuple[int, int], np.dtype[bool]]
+# Define array types
+Array2D: NDArray[Shape["*, *"], Any]  # 2D array of any size of any dtype
+ElevationArray2D: NDArray[Shape["*, *"], float]  # 2D array of any size of dtype float
+BinaryArray2D: NDArray[Shape["*, *"], bool]  # 2D array of any size of dtype bool
 
 # Define functions as interfaces
-## `BinaryClassifierFn`s and `BinaryDenoiserFn`s use the following signature
-## They take a 2D array as input and output a binary 2D array
-## Use partial functions from the `functools` library to create wrapper functions which contain all input arguments other than the input 2D array
-BinaryClassifierFn = Callable[[ElevationArray2D], BinaryArray2D]
-BinaryDenoiserFn = Callable[[BinaryArray2D], BinaryArray2D]
+## `BinaryClassifierFn`s and `BinaryDenoiserFn`s use the following signature:
+### They take a 2D array as input and output a binary 2D array
+### Use partial functions from the `functools` library to create wrapper functions which contain all input arguments other than the input 2D array
+BinaryClassifierFn: Callable[[ElevationArray2D], BinaryArray2D]
+BinaryDenoiserFn: Callable[[BinaryArray2D], BinaryArray2D]
