### 2025-05-10 7e3167e
added explicit type casting
