### 2025-05-10 17b9a21
added explicit type casting and slight refactoring for type errors
