### 2025-05-10 6c757eb
explicit type casting
