### 2025-05-10 015b212
assigned error message to variables before raising error
