### 2025-05-10 7ee58eb
added ruff ignores: T201 for *.ipynb, PD901 and RET504 globally
