+++ pyproject.toml
@@ -159,6 +159,8 @@ extend-select = [
 ignore = [
   "PLR09",    # Too many <...>
   "PLR2004",  # Magic value used in comparison
+  "PD901",    # Generic `df` variable name
+  "RET504",   # unncesarry assignment
 ]
 isort.required-imports = ["from __future__ import annotations"]
 # Uncomment if using a _compat.typing backport
@@ -167,6 +169,7 @@ isort.required-imports = ["from __future__ import annotations"]
 [tool.ruff.lint.per-file-ignores]
 "tests/**" = ["T20"]
 "noxfile.py" = ["T20"]
+"*.ipynb" = ["T201"]
 
 
 [tool.pylint]
