### 2025-05-10 2391133
ruff inline ignore for SIM108, fix for RUF015
