+++ pyproject.toml
@@ -161,6 +161,7 @@ ignore = [
   "PLR2004",  # Magic value used in comparison
   "PD901",    # Generic `df` variable name
   "RET504",   # unncesarry assignment
+  "T201",     # Print statements
 ]
 isort.required-imports = ["from __future__ import annotations"]
 # Uncomment if using a _compat.typing backport
