### 2025-05-10 45c7a9e
added ignore for RUF046 false positive
