### 2025-05-10 e732a4c
assigned error messages to variables for ruff EM101 and EM102
