### 2025-05-10 ae49059
typing.Tuple -> tuple for ruff UP035 and UP006
