### 2025-05-10 9c689ee
removed ambiguous variable name for ruff E741
