### 2025-05-11 ef640be
added explicit TypeAlias and used npt types
