### 2025-05-11 42ac874
fixed logic for determine_ridge_amp()
