### 2025-05-11 7d66062
removed pandas method calls with "inplace=True" ruff:PD002
