### 2025-05-16 293ff43
ignore ruff F722 for numpy shape annotations
