### 2025-05-16 e71fc69
removed redundant inner list() call for ruff C414
