### 2025-05-19 df20764
added ignore for mypy errors concerning bool_mask type
