### 2025-05-19 a2e6a67
added ignore for mypy errors concerning types for optional attributes;
