### 2025-05-19 3b65432
corrected types to handle indefinite number of functions in tuple
