### 2025-05-26 9a4ffe9
added ignore for type error covered by preexisting logical check
