### 2025-05-27 32254d1
changed Callable input to any number of any type of arguments
