### 2025-05-27 4039a26
fixed and ignored pylint errors
