### 2025-05-30 268e88e
added ipykernal for nb support in docs
