+++ pyproject.toml
@@ -54,6 +54,7 @@ docs = [
   "sphinx_autodoc_typehints",
   "furo>=2023.08.17",
   "nbsphinx",
+  "ipykernel",
 ]
 
 [project.urls]
