### 2025-05-30 eb1fc38
added multiple references for intersphinx
