### 2025-06-04 b04252a
added nitpick ignore for nptyping and xarray link for intersphinx
