### 2025-06-04 #7841 @a-vanderheiden
Hi @kthyng, this process is taking longer than expected for me and I wanted to check-in.

I've finished addressing the errors raised by mypy, ruff, and pylint, but am still running into issues building the docs with sphinx without error.

My main source of error with sphinx is coming from the myst extension not being able to find cross reference targets linked in my README.md that are outside of `scrollstats/src` - like the example .ipynb files I've included with scrollstats.

The docs will still build in `scrollstats/docs` and I get nicely formatted pages for the api, but the links in the README "homepage" to files inside the repo but outside of `scrollstats/src` are dead.

I am still actively working on figuring this out and will be reaching out to colleagues who may have experience with this.
