### 2025-08-20 #7841 @a-vanderheiden
@kthyng thanks for checking in! I got majorly stalled on trying to incorporate my python notebooks in the sphinx docs. I still want to use the examples/code in them for the docs, so I’ve decided to just create .md files with similar content for the docs. I'm still working on this.

This will hopefully allow me to get past this step quickly and move on to figuring out the GitHub actions for automated distribution.
