### 2025-09-08 0441576
converted docs from .ipynb to .md; removed parameters.py; updated readme
