+++ README.md
@@ -44,70 +44,37 @@ cd path/to/scrollstats
 conda env create -f environment.yml
 ```
 
-## Using ScrollStats
+## The ScrollStats Workflow
 
-ScrollStats uses interactive jupyter notebooks as the user interface, however
-the underlying [scripting library](src/scrollstats) can also be used on its own.
-Example usage of ScrollStats has been broken up in the three following
-notebooks:
+The ScrollStats Workflow can be broken up into the 3 major steps listed below. Each of these steps is covered in detail in [docs](docs/) with code examples using datasets generated from [a bend on the Lower Brazos River, TX](https://www.google.com/maps/@30.614411,-96.499508,16z).
 
-- [DelineateRidgeAreas.ipynb](DelineateRidgeAreas.ipynb)
-- [CreateVectorDatasets.ipynb](CreateVectorDatasets.ipynb)
-- [CalculateRidgeMetrics.ipynb](CalculateRidgeMetrics.ipynb)
 
-These three notebooks all include detailed instruction on the intended use of
-the ScrollStats library with an included [example dataset](example_data) of a
-bend from the Lower Brazos River, TX. Once you are comfortable using the library
-from the notebooks, feel free to edit the code or make your own scripts to suit
-your needs.
+### 1. [Delineate Ridge Areas](docs/DelineateRidgeAreas.md)
 
-The three notebooks above are written to process one bend at a time. However,
-all of these operations are designed to be easily moidified for batch processing
-of multiple bends, if desired.
+Delineate ridge areas from a DEM to create the ridge area raster. This is achieved by:
+1. applying the profile curvature and residual topography transforms to the
+    DEM
+2. applying a threshold at 0 to these transformed rasters to create binary
+    rasters
+3. finding the union of these binary rasters
+4. denoising the union raster
 
-[CompleteBendProcess.ipynb](CompleteBendProcess.ipynb) contains all of the
-processing steps in the three notebooks above without much annotation so that
-all processes for a bend can be completed at once.
+### 2. [Create Vector Datasets](docs/CreateVectorDatasets.md)
 
-## The ScrollStats workflow
+Create the following vector datasets to define key morphological features of the bend.
+- bend boundary
+- packet boundaries
+- channel centerline
+- ridge lines
+- migration pathways
 
-**1. Set geoprocessing parameters (optional)**
+Some of these datasets are digitized manually, while others are generated by ScrollStats. Details of the vector data creation process can be found in the doc linked above.
 
-- All geoprocessing parameters (such as raster window size) for the example
-  notebooks are kept in [parameters.py](parameters.py) then imported where
-  needed in the notebooks. This allows for a convenient overview of all
-  parameters used in a single run.
-- No part of the internal scripting library is dependent on the values set in
-  [parameters.py](parameters.py).
+### 3. [Calculate Ridge Metrics](docs/CalculateRidgeMetrics.md)
 
-**2. Delineate Ridge Areas**
+Once all of the vector datasets are created and the raster areas are delineated, you can now calculate ridge metrics across the entire bend.
 
-- Delineate ridge areas from a DEM to create the ridge area raster. This is
-  achieved in [DelineateRidgeAreas.ipynb](DelineateRidgeAreas.ipynb) by:
-  1. applying the profile curvature and residual topography transforms to the
-     DEM
-  2. applying a threshold at 0 to these transformed rasters to create binary
-     rasters
-  3. finding the union of these binary rasters
-  4. denoising the union raster
-
-**3. Create Vector Datasets**
-
-- Create the following vector datasets to define key morphological features of
-  the bend. Details of the vector data creation can be found in
[…]
