+++ docs/CalculateRidgeMetrics.md
@@ -1,0 +1,128 @@
+# Calculate Ridge Metrics
+
+This notebook contains the code required to calculate ridge metrics from the previously created datasets:
+- Ridges
+- Migration Pathways
+- Packets
+- Ridge Area Raster
+- DEM
+
+From these datasets, we will calculate ridge width, amplitude, and migration distance at every ridge-transect intersection and plot ridge amplitude on a map.
+
+
+## Output datasets
+
+The `calculate_ridge_metrics` function takes the transects, ridges, clipped ridge area (binary) and DEM rasters as input and returns two GeoDataFrames
+- `rich transects`
+- `itx points`
+
+### Rich Transects
+The `rich_transects` retain the same geometry as the input transects but have the following additional fields:
+- `r_000` ... `r_###`: the distance between the channel centerline and that ridge along the specified transect
+- `dem_signal`: dem values sampled along the transect with 1m sampling 
+- `bin_signal`: binary raster values sampled along the transect with 1m sampling
+- `ridge_count_raster`: the ridge count along the transect calculated from the number of unique ridges in `bin_signal`
+- `fft_spacing`: the fft-derived dominant wavelength from `bin_signal` - corresponds to a representative spacing between ridges along the transect
+
+### Intersection Points
+The `itx` points are the collection of point intersections between the ridges and transects. `itx` contains the  three following ridge metrics in it's attribute field:
+- `pre_mig_dist`: the migration distance from the previous ridge. Also referred to as "ridge spacing"
+- `ridge_width`: the width of the ridge derived from `bin_signal` measured in px
+- `ridge_amp`: the amplitude of the ridge derived from `dem_signal` measured in units of the DEM
+
+`itx` also contains the following fields:
+- `start_distances`: the along-transect distance to the intersection point
+- `transect_position`: vertex position of the intersection point along the transect
+- `metric_confidence`: a simple metric derived from `bin_signal` describing the confidence in other ridge metrics.
+- `relative_vertex_distances`: the relative along-transect distance of the substring that corresponds to the intersection. A relative vertex position of 0.5 is at the exact midpoint of the transect.
+- `vertex_indices`: the indices of the `dem_signal` or `binary_signal` of the transect that correspond to the location of the itx substring
+- `dem_signal`: the section of the transect's `dem_signal` that corresponds to the itx substring
+- `dem_signal_selection`: a subset of `dem_signal` if the intersection point is on the first ridge from the channel
+- `bin_signal`: the section of the transect's `bin_signal` that corresponds to the itx substring
+- `bool_mask`: `bin_signal` expressed as a boolean array
+- `post_mig_dist`: the migration distance to the next ridge
+- `pre_mig_time`: the difference in `deposit_year` between the intersected and previous ridge
+- `post_mig_time`: the difference in `deposit_year` between the intersected and next ridge
+- `pre_mig_rate`: `pre_mig_dist` / `pre_mig_time`
+- `post_mig_rate`: `post_mig_dist` / `post_mig_time`
+- `deposit_year`: `deposit_year` of the corresponding ridge
+- `substring_geometry`: the `LineString` of the substring corresponding to the intersection
+- `geometry`: the `Point` geometry of the intersection
[…]
