+++ docs/index.md
@@ -3,10 +3,22 @@
 ```{toctree}
 :maxdepth: 2
 :hidden:
+:caption: Example Workflow
+
+DelineateRidgeAreas
+CreateVectorDatasets
+CalculateRidgeMetrics
+CompleteScrollStatsWorkflow
+```
+
+```{toctree}
+:maxdepth: 2
+:hidden:
 :caption: API
 
 api/scrollstats
 ```
+
 
 ```{include} ../README.md
 :start-after: <!-- SPHINX-START -->

[…]
