### 2025-09-10 69c4c07
removed readme from docs homepage
