+++ docs/DelineateRidgeAreas.md
@@ -1,18 +1,27 @@
 # Delineate Ridge Areas
-This section demonstrates how ScrollStats delineates ridge areas from an input DEM.
+
+This section demonstrates how ScrollStats delineates ridge areas from an input
+DEM.
 
 Ridge area delineation is accomplished with the following operations:
+
 1. apply profile curvature and residual topography transformations
 2. apply binary classification to the transformed rasters
 3. find the agreement between the binary rasters
 4. denoise the agreement raster
 
 ## Convenience Functions
-The steps above to create what's called the *ridge area raster* are all performed in the two following convenience functions:
+
+The steps above to create what's called the _ridge area raster_ are all
+performed in the two following convenience functions:
+
 - `create_ridge_area_raster`
 - `create_ridge_area_raster_fs`
 
-**`create_ridge_area_raster`** takes a `rasterio.DatasetReader` of the DEM, a shapely `Polygon` of the bend area and any other kwargs as input and returns two `np.ndarray` objects (the clipped ridge area raster and clipped DEM) and the modified `rasterio` metadata to write the arrays to disk if needed
+**`create_ridge_area_raster`** takes a `rasterio.DatasetReader` of the DEM, a
+shapely `Polygon` of the bend area and any other kwargs as input and returns two
+`np.ndarray` objects (the clipped ridge area raster and clipped DEM) and the
+modified `rasterio` metadata to write the arrays to disk if needed
 
 ```python
 import rasterio
@@ -22,8 +31,10 @@ import numpy as np
 from scrollstats import create_ridge_area_raster
 
 # User provided parameters
-RASTER_WINDOW_SIZE = 45 # kernal size for image processing; measured in px
-SMALL_FEATS_SIZE = 500  # all features smaller will be removed in denoising process; measured in px^2
+RASTER_WINDOW_SIZE = 45  # kernel size for image processing; measured in px
+SMALL_FEATS_SIZE = (
+    500  # all features smaller will be removed in denoising process; measured in px^2
+)
 
 # Input Dataset Paths
 dem_path = Path("example_data/input/LBR_025_dem.tif")
@@ -37,16 +48,18 @@ bend = gpd.read_file(bend_path)
 bend_geom = bend.loc[0, "geometry"]
 
 agr_clip, dem_clip, clip_meta = create_ridge_area_raster(
-                                    dem_ds=dem_ds,
-                                    geometry=bend_geom,
-                                    no_data=np.nan,
-                                    window=RASTER_WINDOW_SIZE,
-                                    dx=1,
-                                    small_feats_size=SMALL_FEATS_SIZE
-                                )
+    dem_ds=dem_ds,
+    geometry=bend_geom,
+    no_data=np.nan,
+    window=RASTER_WINDOW_SIZE,
+    dx=1,
+    small_feats_size=SMALL_FEATS_SIZE,
+)
 ```
 
-**`create_ridge_area_raster_fs`** is just a wrapper for `create_ridge_area_raster` taking and returning file paths to act as an interface for the file system
+**`create_ridge_area_raster_fs`** is just a wrapper for
+`create_ridge_area_raster` taking and returning file paths to act as an
+interface for the file system
 
 ```python
 from pathlib import Path
@@ -54,8 +67,10 @@ import numpy as np
 from scrollstats import create_ridge_area_raster_fs
 
 # User provided parameters
-RASTER_WINDOW_SIZE = 45 # kernal size for image processing; measured in px
-SMALL_FEATS_SIZE = 500  # all features smaller will be removed in denoising process; measured in px^2
+RASTER_WINDOW_SIZE = 45  # kernel size for image processing; measured in px
[…]
