+++ README.md
@@ -46,35 +46,44 @@ conda env create -f environment.yml
 
 ## The ScrollStats Workflow
 
-The ScrollStats Workflow can be broken up into the 3 major steps listed below. Each of these steps is covered in detail in [docs](docs/) with code examples using datasets generated from [a bend on the Lower Brazos River, TX](https://www.google.com/maps/@30.614411,-96.499508,16z).
-
+The ScrollStats Workflow can be broken up into the 3 major steps listed below.
+Each of these steps is covered in detail in [docs](docs/) with code examples
+using datasets generated from
+[a bend on the Lower Brazos River, TX](https://www.google.com/maps/@30.614411,-96.499508,16z).
 
 ### 1. [Delineate Ridge Areas](docs/DelineateRidgeAreas.md)
 
-Delineate ridge areas from a DEM to create the ridge area raster. This is achieved by:
-1. applying the profile curvature and residual topography transforms to the
-    DEM
+Delineate ridge areas from a DEM to create the ridge area raster. This is
+achieved by:
+
+1. applying the profile curvature and residual topography transforms to the DEM
 2. applying a threshold at 0 to these transformed rasters to create binary
-    rasters
+   rasters
 3. finding the union of these binary rasters
 4. denoising the union raster
 
 ### 2. [Create Vector Datasets](docs/CreateVectorDatasets.md)
 
-Create the following vector datasets to define key morphological features of the bend.
+Create the following vector datasets to define key morphological features of the
+bend.
+
 - bend boundary
 - packet boundaries
 - channel centerline
 - ridge lines
 - migration pathways
 
-Some of these datasets are digitized manually, while others are generated by ScrollStats. Details of the vector data creation process can be found in the doc linked above.
+Some of these datasets are digitized manually, while others are generated by
+ScrollStats. Details of the vector data creation process can be found in the doc
+linked above.
 
 ### 3. [Calculate Ridge Metrics](docs/CalculateRidgeMetrics.md)
 
-Once all of the vector datasets are created and the raster areas are delineated, you can now calculate ridge metrics across the entire bend.
+Once all of the vector datasets are created and the raster areas are delineated,
+you can now calculate ridge metrics across the entire bend.
 
-These metrics include ridge amplitude, width, and migration distance for every intersection of a ridge and migration pathway.
+These metrics include ridge amplitude, width, and migration distance for every
+intersection of a ridge and migration pathway.
 
 ## Contributing
 
