+++ docs/CreateVectorDatasets.md
@@ -1,57 +1,139 @@
 # Create Vector Datasets
 
-This notebook contains general instructions for the creation of the vector data sets required by ScrollStats. 
+This notebook contains general instructions for the creation of the vector data
+sets required by ScrollStats.
 
-An example dataset of a bend from the Lower Brazos River, Texas, is included in this notebook to provide a visual aid to the user when creating datasets for their own bend. 
+An example dataset of a bend from the Lower Brazos River, Texas, is included in
+this notebook to provide a visual aid to the user when creating datasets for
+their own bend.
 
-Some datasets need to be manually created while others are programmatically generated from the manually created datasets. 
+Some datasets need to be manually created while others are programmatically
+generated from the manually created datasets.
 
 ---
 
-ScrollStats uses the following vector datasets for a bend in order to calculate ridge metrics:
+ScrollStats uses the following vector datasets for a bend in order to calculate
+ridge metrics:
+
 - **Channel Centerline (LineString):**
-    - The channel centerline polyline should not intersect the bend boundary polygon at all and should extend past inner bank of the bend boundary by 1-2 channel widths.
-- **Bend Boundary (Polygon):** 
-    - The bend boundary in this case defines the boundary encompasses the raised platform of ridge and swale topography between the active channel and the relatively smooth ancestral floodplain. Each bend boundary should have a corresponding bend_id that can be used to relate data together. For example, the 25th bend on the Lower Brazos River was given the beind_id `LBR_025`.
+  - The channel centerline polyline should not intersect the bend boundary
+    polygon at all and should extend past inner bank of the bend boundary by 1-2
+    channel widths.
+- **Bend Boundary (Polygon):**
+  - The bend boundary in this case defines the boundary encompasses the raised
+    platform of ridge and swale topography between the active channel and the
+    relatively smooth ancestral floodplain. Each bend boundary should have a
+    corresponding bend_id that can be used to relate data together. For example,
+    the 25th bend on the Lower Brazos River was given the beind_id `LBR_025`.
 - **Packet Boundaries (Polygon):**
-    - The packet boundaries are polygons that fit perfectly within, and cover entirely, the bend boundary polygons. Packet boundaries encompass groups of ridges with similar trajectories. Packet boundaries should have `bend_id` column as well as a `packet_id` column. The `bend_id` can be used as a foreign key to relate the packets to their bend and the simple `packet_id` (ex. `p_01`) can be used to diffferentiate the packets within each bend. There is no guarantee of an inherent order with packets, but in general, they can and should be numbered incrementally from the most ancestral to the most recent.
+  - The packet boundaries are polygons that fit perfectly within, and cover
+    entirely, the bend boundary polygons. Packet boundaries encompass groups of
+    ridges with similar trajectories. Packet boundaries should have `bend_id`
+    column as well as a `packet_id` column. The `bend_id` can be used as a
+    foreign key to relate the packets to their bend and the simple `packet_id`
+    (ex. `p_01`) can be used to diffferentiate the packets within each bend.
+    There is no guarantee of an inherent order with packets, but in general,
+    they can and should be numbered incrementally from the most ancestral to the
+    most recent.
 - **Ridges (LineString):**
[…]

[…]
