### 2025-09-16 0943ec0
added nox to dev dependencies
