### 2025-09-17 5aa5cf9
added numpydoc docstrings to transect.py
