+++ src/scrollstats/transecting/transect.py
@@ -15,6 +15,19 @@ from tqdm import tqdm
 
 ## Simple Geometric functions: accepts and returns arrays of coordinates
 def curvature(line: LineString) -> tuple[np.ndarray[float], np.ndarray[float]]:
+    """
+    Calculates curvature and angle (alpha) along a given line.
+
+    Parameters
+    ----------
+    line : LineString
+        The line for which to calculate curvature and angle.
+
+    Returns
+    -------
+    tuple[np.ndarray[float], np.ndarray[float]]
+        A tuple containing two numpy arrays: curvature and alpha
+    """
     x, y = line.xy
 
     # Get dx, dy
@@ -48,7 +61,19 @@ def curvature(line: LineString) -> tuple[np.ndarray[float], np.ndarray[float]]:
 
 
 def calc_dist_along_line(line: LineString) -> np.ndarray[float]:
-    """Calculate an array of distance along a given line"""
+    """
+    Calculate an array of distance along a given line
+
+    Parameters
+    ----------
+    line : LineString
+        The line for which to calculate distance along.
+
+    Returns
+    -------
+    np.ndarray[float]
+        Array of distance values for each vertex.
+    """
 
     x, y = line.xy
 
@@ -61,6 +86,21 @@ def calc_dist_along_line(line: LineString) -> np.ndarray[float]:
 
 
 def calc_dist(p1: np.ndarray[float], p2: np.ndarray[float]) -> np.ndarray[float]:
+    """
+    Calculates the distance element-wise between two arrays of points
+
+    Parameters
+    ----------
+    p1 : np.ndarray[float]
+        First array of points.
+    p2 : np.ndarray[float]
+        Second array of points.
+
+    Returns
+    -------
+    np.ndarray[float]
+        Array of distances between corresponding points in p1 and p2.
+    """
     # p1 and p2 are both (n,2) arrays of coordinates
     return np.sqrt((p1[:, 0] - p2[:, 0]) ** 2 + (p1[:, 1] - p2[:, 1]) ** 2)
 
@@ -69,9 +109,23 @@ def law_of_cos(
     a: np.ndarray[float], b: np.ndarray[float], c: np.ndarray[float]
 ) -> np.ndarray[float]:
     """
-    This function uses law of cosines to calc the interior angle formed by two connected lines
+    This function uses law of cosines to calculate the interior angle formed by two connected lines
     Traveling down the line, vertices are a, b, c (a=beg, b=mid, c=end)
     Distances between vertices are represented with '_' ex. a_b is the distance between a and b
+
+    Parameters
+    ----------
+    a : np.ndarray[float]
+        Array of coordinates for point a.
+    b : np.ndarray[float]
+        Array of coordinates for point b.
+    c : np.ndarray[float]
+        Array of coordinates for point c.
+
+    Returns
+    -------
+    np.ndarray[float]
+        Array of interior angles at point b formed by lines ab and bc.
     """
 
     # Calculate distances between points
@@ -86,7 +140,21 @@ def law_of_cos(
 
 
 def calc_dxdy(angle: np.ndarray[float], length: float = 100) -> np.ndarray[float]:
-    """Calculates the delta x and y given an angle and a length"""
+    """
+    Calculates the delta x and y given an angle and a length
+
+    Parameters
+    ----------
+    angle : np.ndarray[float]
+        Angle(s) in radians.
+    length : float, optional
+        Length of the vector, by default 100.
+
+    Returns
+    -------
+    np.ndarray[float]
+        A 2D array where the first row contains delta x values and the second row contains delta y values.
+    """
 
     dx = np.cos(angle) * length
     dy = np.sin(angle) * length
@@ -102,10 +170,21 @@ def add_sub_90(
[…]
