### 2025-09-22 df491cd
imported ScrollStats in conf.py
