### 2025-09-27 1a9d61f
updated the Getting Started instructions and removed mentions of conda
