+++ environment.yml
@@ -1,15 +1,0 @@
-name: scroll
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - python=3.10
-  - notebook
-  - geopandas
-  - matplotlib
-  - scipy
-  - rasterio
-  - numba
-  - pytest
-  - tqdm
-  - nox
