+++ docs/CreateVectorDatasets.md
@@ -132,7 +132,8 @@ combination of the bend_id and packet_id columns forms the primary key for the
 packets.
 
 Example bend and packet boundaries can be found in the
-[example_data](https://github.com/tamu-edu/scrollstats/tree/main/example_data) directory included with ScrollStats. These boundaries are mapped below.
+[example_data](https://github.com/tamu-edu/scrollstats/tree/main/example_data)
+directory included with ScrollStats. These boundaries are mapped below.
 
 ```python
 from __future__ import annotations
@@ -183,7 +184,8 @@ After the ridge area rasters are created, manually create the ridge lines and
 channel centerlines.
 
 Example centerline and ridgelines can be found in the
-[example_data](https://github.com/tamu-edu/scrollstats/tree/main/example_data) directory included with ScrollStats. The centerline and ridgelines are mapped
+[example_data](https://github.com/tamu-edu/scrollstats/tree/main/example_data)
+directory included with ScrollStats. The centerline and ridgelines are mapped
 below.
 
 ```python
