### 2025-09-27 c2a8875
changed validate-pyproject to pinned version of validate-pyproject-schema-store. Releases of schema-store after 2025.08.29 start failing
