+++ .pre-commit-config.yaml
@@ -77,11 +77,10 @@ repos:
         entry: PyBind|Numpy|Cmake|CCache|Github|PyTest
         exclude: .pre-commit-config.yaml
 
-  - repo: https://github.com/abravalheri/validate-pyproject
-    rev: "v0.24.1"
+  - repo: https://github.com/henryiii/validate-pyproject-schema-store
+    rev: "2025.08.29"
     hooks:
       - id: validate-pyproject
-        additional_dependencies: ["validate-pyproject-schema-store[all]"]
 
   - repo: https://github.com/python-jsonschema/check-jsonschema
     rev: "0.33.0"
