### 2025-09-29 f8b51ab
added new ignore for pylint
