+++ pyproject.toml
@@ -186,6 +186,7 @@ messages_control.disable = [
   "missing-module-docstring",
   "missing-function-docstring",
   "wrong-import-position",
+  "too-many-lines",
 ]
 
 [tool.codespell]
