### 2025-09-29 3dc4d44
updated py versions in CI workflow
