+++ .github/workflows/ci.yml
@@ -40,7 +40,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ["3.9", "3.13"]
+        python-version: ["3.10", "3.11", "3.12"]
         runs-on: [ubuntu-latest, windows-latest, macos-14]
 
         include:
