### 2025-09-30 384a428
removed windows os and pypy from CI
