+++ .github/workflows/ci.yml
@@ -41,11 +41,7 @@ jobs:
       fail-fast: false
       matrix:
         python-version: ["3.10", "3.11", "3.12"]
-        runs-on: [ubuntu-latest, windows-latest, macos-14]
-
-        include:
-          - python-version: "pypy-3.10"
-            runs-on: ubuntu-latest
+        runs-on: [ubuntu-latest, macos-14]
 
     steps:
       - uses: actions/checkout@v4
