### 2025-09-30 1383a7e
removed conda based CI
