+++ .github/workflows/main.yml
@@ -1,47 +1,0 @@
-name: ScrollStats Test
-
-# Controls when the workflow will run
-on:
-  # Triggers the workflow on push or pull request events but only for the "main" branch
-  push:
-    branches: ["main"]
-  pull_request:
-    branches: ["main"]
-
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
-jobs:
-  # This workflow contains a single job called "test"
-  test:
-    # The type of runner that the job will run on
-    name: Build OS ${{ matrix.os }}
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        # TODO: add "windows-latest"
-        os: ["ubuntu-latest", "macos-latest"]
-
-    # Steps represent a sequence of tasks that will be executed as part of the job
-    steps:
-      # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
-      - uses: actions/checkout@v4
-      - uses: conda-incubator/setup-miniconda@v3
-        with:
-          activate-environment: scroll
-          environment-file: environment.yml
-          miniforge-version: "latest"
-
-      # Check the conda info to make sure it was initialized correctly
-      - name: Check Conda info
-        # You MUST specify the shell for conda commands to work. See https://github.com/conda-incubator/setup-miniconda?tab=readme-ov-file#important
-        shell: bash -el {0}
-        run: |
-          conda info
-          conda list
-
-      - name: Run tests
-        shell: bash -el {0}
-        run: |
-          pytest test_all.py
