### 2025-11-10 a45e68a
added api docs to build job
