### 2025-11-10 fd2408d
added sphinx.ext.apidoc
