### 2025-11-16 0c2db7e
docs build job without sphinx-apidocs
