+++ .readthedocs.yaml
@@ -7,11 +7,16 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.12"
-  commands:
-    - asdf plugin add uv
-    - asdf install uv latest
-    - asdf global uv latest
-    - uv venv
-    - uv pip install .[docs]
-    - .venv/bin/python -m sphinx -T -b html -d docs/_build/doctrees -D
-      language=en docs $READTHEDOCS_OUTPUT/html
+  jobs:
+    post_checkout:
+      - git fetch --tags
+    install:
+      - asdf plugin add uv
+      - asdf install uv latest
+      - asdf global uv latest
+      - uv venv
+      - uv pip install .[docs]
+    build:
+      html:
+        - .venv/bin/python -m sphinx -T -b html -d docs/_build/doctrees -D
+          language=en docs $READTHEDOCS_OUTPUT/html
