### 2025-11-16 857132d
changed order of sphinx commands
