### 2025-11-16 1570659
updated python execution for rtd to `uv run python`
