### 2025-11-16 3f6e5db
updated apidoc exec to `.venv/bin/sphinx-apidoc`
