### 2025-11-16 996824d
fixed error with calling sphinx-apidoc executable
