### 2025-11-16 #7841 @a-vanderheiden
> Hi [@a-vanderheiden](https://github.com/a-vanderheiden) , thanks for your submission. I’ll be taking over from here. Adithi

Hi @adithirgis, thanks for taking this up!

Just so you know, I released v0.1.2 last night just before you were assigned as editor. It's mostly just contains changes I made to .readthedocs.yaml to add the api docs and address version issues I was having with the docs.
