### 2025-11-17 5c8d0b0
add lazy-loader to dependencies
