### 2025-11-17 274e120
replace direct imports with lazy imports
