### 2025-11-17 caaec78
replace direct imports with lazy imports
