+++ src/scrollstats/delineation/__init__.py
@@ -1,13 +1,16 @@
 from __future__ import annotations
 
-__all__ = [
-    "LineSmoother",
-    "create_ridge_area_raster",
-    "create_ridge_area_raster_fs",
-    "quadratic_profile_curvature",
-    "residual_topography",
-]
+import lazy_loader as lazy
 
-from .line_smoother import LineSmoother
-from .raster_classifiers import quadratic_profile_curvature, residual_topography
-from .ridge_area_raster import create_ridge_area_raster, create_ridge_area_raster_fs
+__getattr__, __dir__, __all__ = lazy.attach(
+    __name__,
+    submodules=[],
+    submod_attrs={
+        "line_smoother": ["LineSmoother"],
+        "raster_classifiers": ["quadratic_profile_curvature", "residual_topography"],
+        "ridge_area_raster": [
+            "create_ridge_area_raster",
+            "create_ridge_area_raster_fs",
+        ],
+    },
+)
