### 2025-11-17 880d6d2
added lazy_loader to mypy ignore list
