### 2025-11-17 b266e33
added lazy_loader to mypy ignore list
