### 2026-01-06 #7841 @martinvonk
> 👋 [@martinvonk](https://github.com/martinvonk) - Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Thank you for asking. I’ll have to decline though, as this isn’t my area of expertise.
