### 2026-01-11 tamu-edu/scrollstats#31 @Atreyagaurav
[JOSS] Add Code Examples
Add examples that users can run. Alongside the `example_data`  in the repo, adding example code that users can just run to generate the outputs might be better, compared to having the data here and then the code in the documentation (which should be linked in the README and repository metadata).

I recommend one python file per one step described in the README, with a clear description of required input files and output files generated from each step.

Related: https://github.com/openjournals/joss-reviews/issues/9736
