### 2026-01-15 tamu-edu/scrollstats#31 @a-vanderheiden
I think that's a good idea. I'll take from the examples in the docs to make these python files and save them in a folder called `scripts`.

I've also linked the documentation in the README on branch [joss-iss30](https://github.com/tamu-edu/scrollstats/tree/joss-iss30) as well the repo metadata.
