### 2026-01-20 tamu-edu/scrollstats#32 @Atreyagaurav
[JOSS] Spelling mistakes in documentation
Minor comments.

Related: https://github.com/openjournals/joss-reviews/issues/9736

Spelling mistakes in the documentation `docs/index.md`.

> These migration pathways are then used to sample a DEM and other derived
> **topgraphic** raster data to create a series of 1 dimensional signals from which
> the ridge morphometrics of amplitude, width, and spacing can be calculated for
> every intersection of a ridge and migration pathway.

topographic

> Figure 1: Measures of ridge amplitude (orange), width (purple), and spacing (green) are shown at the intersection, ridge, and migration pathway scales. Aggregate values represent the **medain** value of each measurement taken at a ridge or migration pathway

median

Please check any other places as well.
