### 2026-01-20 tamu-edu/scrollstats#33 @Atreyagaurav
[JOSS] Code in documentation does not run
Related: https://github.com/openjournals/joss-reviews/issues/9736

The code in the first segment of the documentation does not run.

Documentation code from: https://scrollstats.readthedocs.io/en/stable/DelineateRidgeAreas.html
```python
import rasterio
import geopandas as gpd
import numpy as np

from scrollstats import create_ridge_area_raster

# User provided parameters
RASTER_WINDOW_SIZE = 45  # kernel size for image processing; measured in px
SMALL_FEATS_SIZE = (
    500  # all features smaller will be removed in denoising process; measured in px^2
)

# Input Dataset Paths
dem_path = Path("example_data/input/LBR_025_dem.tif")
bend_path = Path("example_data/input/LBR_025_bend.geojson")

# Read in the DEM
dem_ds = rasterio.open(dem_path)

# Read in the bend area as a GeoDataFrame; then get the Polygon geometry
bend = gpd.read_file(bend_path)
bend_geom = bend.loc[0, "geometry"]

agr_clip, dem_clip, clip_meta = create_ridge_area_raster(
    dem_ds=dem_ds,
    geometry=bend_geom,
    no_data=np.nan,
    window=RASTER_WINDOW_SIZE,
    dx=1,
    small_feats_size=SMALL_FEATS_SIZE,
)
```

The code uses `Path` but it is not imported: `NameError: name 'Path' is not defined`

`from pathlib import Path` line from the second code block should be included in the first one.
