### 2026-01-20 tamu-edu/scrollstats#34 @Atreyagaurav
[JOSS] Different results than shown in the documentation
Related: https://github.com/openjournals/joss-reviews/issues/9736

While running the example code from: https://scrollstats.readthedocs.io/en/stable/DelineateRidgeAreas.html

The plots when running the code, using the provided example data, is different from the one in the documentation:

Image from documentation:

Output of the code:

Similarly, other outputs also look little different:

I would suggest generating the outputs for the documentation from the repository automatically (through CI/CD), or at least as a part of the commit/release process to avoid discrepancies like this.
