### 2026-01-22 tamu-edu/scrollstats#32 @a-vanderheiden
Thank you for pointing this out! I'll get these typos fixed and let you know.
