### 2026-01-22 tamu-edu/scrollstats#34 @a-vanderheiden
This is interesting, I'll have to take a closer look at this and see what's causing the discrepancy. Either way, generating the images as a CI/CD step before the docs are built is a good idea.  I'll try implementing this
