### 2026-01-26 6754eec
created example scripts demonstrating scrollstats use
