### 2026-01-26 tamu-edu/scrollstats#31 @a-vanderheiden
I just pushed a branch ([joss-iss31](https://github.com/tamu-edu/scrollstats/tree/joss-iss31)) with these example [scripts](https://github.com/tamu-edu/scrollstats/tree/joss-iss31/scripts). Is this what you had in mind?

For the most part, they follow the code that is present in the docs, but they have better description of the input/output files and parameters as well as updated plots.
