### 2026-01-26 #9736 @a-vanderheiden
Hi @adithirgis, thanks for checking-in. So far the review has been going well. I've been in the process of packing and moving apartments this week, so I have not had as much time to address issues raised by the reviewers as I had hoped. I was able to take some time this afternoon/evening though to work on [issue 31](https://github.com/tamu-edu/scrollstats/issues/31).
