### 2026-02-02 a2e9fb6
generate plots for docs with CI/CD
