+++ scripts/plots/delineate_ridge_areas.py
@@ -1,0 +1,201 @@
+# Generates plots for docs
+
+from functools import partial
+from pathlib import Path
+
+import geopandas as gpd
+import matplotlib.pyplot as plt
+import numpy as np
+import rasterio
+from matplotlib.colors import BoundaryNorm, ListedColormap
+from scipy import ndimage
+
+from scrollstats.delineation.raster_classifiers import (
+    quadratic_profile_curvature,
+    residual_topography,
+)
+from scrollstats.delineation.raster_denoisers import remove_small_feats_w_flip
+from scrollstats.delineation.ridge_area_raster import clip_raster
+
+
+
+# Input Paths
+dem_path = Path("example_data/input/LBR_025_dem.tif")
+bend_path = Path("example_data/input/LBR_025_bend.geojson")
+
+# Output Directory
+output_dir = Path("example_data/output")
+
+# Image variables
+img_dir = Path("img")
+dpi = 100
+
+
+# User provided parameters
+RASTER_WINDOW_SIZE = 45  # kernel size for image processing; measured in px
+SMALL_FEATS_SIZE = (
+    500  # all features smaller will be removed in denoising process; measured in px^2
+)
+
+# Open the DEM dataset
+dem_handle = rasterio.open(dem_path)
+dem = dem_handle.read(1)
+
+# Find no data pixels
+if np.isnan(dem_handle.nodata):
+    no_data_mask = np.isnan(dem_handle.read(1))
+else:
+    no_data_mask = dem == dem_handle.nodata
+
+# Set no data values to absurd integer to avoid errors from the classifier functions with infinite values
+dem[no_data_mask] = -99999
+
+# Calculate the transformation rasters
+#profc = quadratic_profile_curvature(dem, RASTER_WINDOW_SIZE, dx=1)
+#rt = residual_topography(dem, RASTER_WINDOW_SIZE)
+profc_cache = "profc.npy"
+rt_cache = "rt.npy"
+
+try:
+    profc = np.load(profc_cache)
+    rt = np.load(rt_cache)
+except FileNotFoundError:
+    profc = quadratic_profile_curvature(dem, RASTER_WINDOW_SIZE, dx=1)
+    np.save(profc_cache, profc)
+
+    rt = residual_topography(dem, RASTER_WINDOW_SIZE)
+    np.save(rt_cache, rt)
+
+
+# Plot
+fig, (ax1, ax2, ax3) = plt.subplots(1, 3, figsize=(15, 4))
+
+mapper = ax1.imshow(profc, vmin=-0.05, vmax=0.05, cmap="seismic")
+fig.colorbar(mapper, ax=ax1)
+ax1.set_axis_off()
+ax1.set_title("Profile Curvature")
+
+mapper = ax2.imshow(dem, vmin=60)
+fig.colorbar(mapper, ax=ax2)
+ax2.set_axis_off()
+ax2.set_title("Original DEM")
+
+mapper = ax3.imshow(rt, vmin=-2, vmax=2, cmap="seismic")
+fig.colorbar(mapper, ax=ax3)
+ax3.set_axis_off()
+ax3.set_title("Residual Topography")
+
+plt.tight_layout()
+plt.savefig(img_dir/"pc_v_dem_v_rt.png", dpi=dpi)
+
+
+
+# classify each transformed raster from step 1
+profc_bc = profc > 0
+rt_bc = rt > 0
+
+# Plot the classification results
+## 00 - no ridge
[…]
