### 2026-02-02 bc999ac
updated DelineateRidgeAreas.md to match updated plots
