+++ docs/DelineateRidgeAreas.md
@@ -34,9 +34,7 @@ from scrollstats import create_ridge_area_raster
 
 # User provided parameters
 RASTER_WINDOW_SIZE = 45  # kernel size for image processing; measured in px
-SMALL_FEATS_SIZE = (
-    500  # all features smaller will be removed in denoising process; measured in px^2
-)
+SMALL_FEATS_SIZE = 500   # all features smaller will be removed in denoising process; measured in px^2
 
 # Input Dataset Paths
 dem_path = Path("example_data/input/LBR_025_dem.tif")
@@ -70,9 +68,7 @@ from scrollstats import create_ridge_area_raster_fs
 
 # User provided parameters
 RASTER_WINDOW_SIZE = 45  # kernel size for image processing; measured in px
-SMALL_FEATS_SIZE = (
-    500  # all features smaller will be removed in denoising process; measured in px^2
-)
+SMALL_FEATS_SIZE = 500   # all features smaller will be removed in denoising process; measured in px^2
 
 # Set paths to datasets on the file system
 dem_path = Path("example_data/input/LBR_025_dem.tif")
@@ -109,6 +105,8 @@ the union of these two binary rasters represent the true ridge locations within
 the DEM.
 
 ```python
+# Generates plots for docs
+
 from functools import partial
 from pathlib import Path
 
@@ -119,10 +117,6 @@ import rasterio
 from matplotlib.colors import BoundaryNorm, ListedColormap
 from scipy import ndimage
 
-from scrollstats.delineation import (
-    create_ridge_area_raster,
-    create_ridge_area_raster_fs,
-)
 from scrollstats.delineation.raster_classifiers import (
     quadratic_profile_curvature,
     residual_topography,
@@ -139,9 +133,7 @@ output_dir = Path("example_data/output")
 
 # User provided parameters
 RASTER_WINDOW_SIZE = 45  # kernel size for image processing; measured in px
-SMALL_FEATS_SIZE = (
-    500  # all features smaller will be removed in denoising process; measured in px^2
-)
+SMALL_FEATS_SIZE = 500   # all features smaller will be removed in denoising process; measured in px^2
 ```
 
 ```python
@@ -163,7 +155,7 @@ profc = quadratic_profile_curvature(dem, RASTER_WINDOW_SIZE, dx=1)
 rt = residual_topography(dem, RASTER_WINDOW_SIZE)
 
 # Plot
-fig, (ax1, ax2, ax3) = plt.subplots(1, 3, figsize=(20, 5))
+fig, (ax1, ax2, ax3) = plt.subplots(1, 3, figsize=(15, 4))
 
 mapper = ax1.imshow(profc, vmin=-0.05, vmax=0.05, cmap="seismic")
 fig.colorbar(mapper, ax=ax1)
@@ -179,6 +171,8 @@ mapper = ax3.imshow(rt, vmin=-2, vmax=2, cmap="seismic")
 fig.colorbar(mapper, ax=ax3)
 ax3.set_axis_off()
 ax3.set_title("Residual Topography")
+
+plt.tight_layout()
 ```
 
 ![pc_v_dem_v_rt.png](../img/pc_v_dem_v_rt.png)
@@ -207,7 +201,7 @@ rt_vis[rt_bc] = 10
 
 comp_vis = profc_vis + rt_vis
 
-fig, ax = plt.subplots(1, 1, figsize=(7, 5))
[…]
