### 2026-02-02 611f09c
added command to build plots for delineation docs
