### 2026-02-02 913b7e2
removed local array cache step
