+++ scripts/delineate_ridge_area_raster.py
@@ -1,15 +1,17 @@
+from __future__ import annotations
+
 from pathlib import Path
 
-import rasterio
 import geopandas as gpd
-import numpy as np
 import matplotlib.pyplot as plt
+import numpy as np
+import rasterio
 
 from scrollstats import create_ridge_area_raster
 
 # User provided parameters
 RASTER_WINDOW_SIZE = 45  # kernel width for image processing; measured in px
-SMALL_FEATS_SIZE = 500   # minimum feature size for image classification; all features smaller will be removed in denoising process; measured in px^2
+SMALL_FEATS_SIZE = 500  # minimum feature size for image classification; all features smaller will be removed in denoising process; measured in px^2
 
 
 # Input Dataset Paths
@@ -28,19 +30,19 @@ dem_ds = rasterio.open(dem_path)
 bend = gpd.read_file(bend_path)
 bend_geom = bend.loc[0, "geometry"]
 
-# Delineate ridge areas 
+# Delineate ridge areas
 ## This function applies two independent classification functions to the DEM to identify ridge areas: profile curvature and residual topography
 ## Both of these functions return 2D float arrays where values greater 0 than indicate the presense of a ridge. A threshold of 0 is applied to both of these float arrays to create binary arrays.
 ## The union of these two binary arrays is then subject to a denoising process to create the ridge area raster.
 ## `create_ridge_area_raster` returns the ridge area raster and the clipped DEM as numpy arrays along with the required metadata to write them to disk as tifs with rasterio
 ridge_area_raster, dem_clip, clip_meta = create_ridge_area_raster(
-    dem_ds=dem_ds,                        # input DEM; rasterio.DatasetReader not np array
-    geometry=bend_geom,                   # bend polygon containing ridge & swale topography
-                                          # Set kwargs for raster delineation functions
-    no_data=np.nan,                       ## Set no data value for the clipped raster
-    window=RASTER_WINDOW_SIZE,            ## Set kernel size for localized operations (profile curvature & residual topography)
-    dx=1,                                 ## Set grid spacing of raster
-    small_feats_size=SMALL_FEATS_SIZE,    ## Set min size of image objects to be preserved in image denoising 
+    dem_ds=dem_ds,  # input DEM; rasterio.DatasetReader not np array
+    geometry=bend_geom,  # bend polygon containing ridge & swale topography
+    # Set kwargs for raster delineation functions
+    no_data=np.nan,  ## Set no data value for the clipped raster
+    window=RASTER_WINDOW_SIZE,  ## Set kernel size for localized operations (profile curvature & residual topography)
+    dx=1,  ## Set grid spacing of raster
+    small_feats_size=SMALL_FEATS_SIZE,  ## Set min size of image objects to be preserved in image denoising
 )
 
 # Write arrays to disk
@@ -64,14 +66,14 @@ affine = clip_meta.get("transform")
 ul_clip_geox = affine.xoff
 ul_clip_geoy = affine.yoff
 
-### Get img coordinates of DEM array corresponding to ridge area raster upperleft corner 
+### Get img coordinates of DEM array corresponding to ridge area raster upperleft corner
 ul_clip_imgx, ul_clip_imgy = ~dem_ds.transform * (ul_clip_geox, ul_clip_geoy)
 
 ## Prepare DEM and ridge area raster for the plot
 ### Clip DEM to ridge area raster envelope
 dem_vis = dem_ds.read(1)[
-    int(ul_clip_imgy): int(ul_clip_imgy) + ridge_area_raster.shape[0], 
[…]
