### 2026-02-02 a96ca81
moved plot build step to html build
