### 2026-02-02 tamu-edu/scrollstats#34 @a-vanderheiden
@Atreyagaurav I looked into this issue this weekend and I think these discrepancies were the result of me saving plots from previous versions of scrollstats.

I've implemented your suggestion of generating the plots as part of the CI/CD and pushed it to the branch [joss-review](https://github.com/tamu-edu/scrollstats/tree/joss-review). You can see the results on the [Delineate Ridge Areas](https://scrollstats.readthedocs.io/en/joss-review/DelineateRidgeAreas.html) page of the joss-review version of the docs.

**Implementation**
I created a subdirectory in scripts called [plots](https://github.com/tamu-edu/scrollstats/tree/joss-review/scripts/plots) where I'll keep the scripts to create plots used in the docs. Executing this script was then added to the docs session in [noxfile.py](https://github.com/tamu-edu/scrollstats/blob/a96ca815c94a188c9c08528af6240cd67f0d8cc0/noxfile.py#L73) and to [.readthedocs.yml](https://github.com/tamu-edu/scrollstats/blob/a96ca815c94a188c9c08528af6240cd67f0d8cc0/.readthedocs.yaml#L21) to build locally and remotely. Is there another way you would recommend implementing this? This implementation seems to work well enough, but I'm open to changing this if there is a better/more commonly used pattern.

I'll continue working on implementing the same solution for the other plots used in the docs, though I don't think there are any other major discrepancies with the other plots.
