### 2026-02-02 tamu-edu/scrollstats#32 @a-vanderheiden
These typos (and others) have been fixed in commit e34c64436af32c5258f9d039b949d6c8e24117b1.
