### 2026-02-02 #9736 @a-vanderheiden
@changliao1025 @Atreyagaurav I've decided to commit all changes from the review process to the branch [joss-review](https://github.com/tamu-edu/scrollstats/tree/joss-review). If needed, I'll create/publish branches off of joss-review, but otherwise I'll just commit directly to this branch. Once the review is complete, I'll merge all of the these changes in joss-review with main and create a new release.

I've incorporated the commits made previously on [joss-iss30](https://github.com/tamu-edu/scrollstats/tree/joss-iss30) and [joss-iss31](https://github.com/tamu-edu/scrollstats/tree/joss-iss31) into this branch as well. Are yall ok if I move forward with deleting those two issue branches?
