### 2026-02-02 tamu-edu/scrollstats#34 @Atreyagaurav
Hi,

I don't see the changes in the documentation website now. I guess it comes up when you make a release?

As long as it is automated, through `.readthedocs.yml` or github CI/CD, it is fine. This helps keep the documentation consistent and saves the time of the user (who might go "what did I do wrong?" because they can not replicate the results).
