### 2026-02-02 tamu-edu/scrollstats#33 @Atreyagaurav
Great

Aside: Can you open a pull request for `joss-review` to `main` so that we can link all these issues there?
