### 2026-02-03 tamu-edu/scrollstats#34 @a-vanderheiden
The changes aren't live on the stable or latest version of the docs, but I do have changes published on the joss-review version at https://scrollstats.readthedocs.io/en/joss-review/DelineateRidgeAreas.html. Are you able to see them there?

You're correct though that with my current CD process there wont be a change to the latest or stable version of the docs until I create a new release. My plan is to commit all of the changes from the review process to the joss-review branch, then merge to main and create a single release at the end.

> As long as it is automated, through .readthedocs.yml or github CI/CD, it is fine. This helps keep the documentation consistent and saves the time of the user (who might go "what did I do wrong?" because they can not replicate the results).

Sounds good! I'll keep it how it is then. And I agree, it makes the user experience with the docs much better. Thank you for the suggestion!
