### 2026-02-06 bb930c4
added script to create plots for vector data docs
