### 2026-02-06 418915f
updated vector docs code to match autogenerated output
