### 2026-02-06 73e8bdb
update nox docs step with vector data plots
