### 2026-02-06 a516397
update readthedocs with vector data plots
