+++ .readthedocs.yaml
@@ -19,6 +19,7 @@ build:
     build:
       html:
         - uv run python scripts/plots/delineate_ridge_areas.py
+        - uv run python scripts/plots/create_vector_datasets.py
         - .venv/bin/sphinx-apidoc -o docs/api --module-first --no-toc --force
           src/scrollstats
         - uv run python -m sphinx -T -b html -d docs/_build/doctrees -D
