### 2026-02-10 #9736 @Atreyagaurav
Hi @adithirgis , thank you for checking in. The author has not made any updates on the manuscript after my initial response about the required sections not being present in the `paper.md` file. So I am waiting on that.
