### 2026-02-10 #9736 @Atreyagaurav
> Hi [@Atreyagaurav](https://github.com/Atreyagaurav), thank you for the early feedback!
>
> You're correct; the paper requirements have changed since my original submission. I'll work on adding those missing sections and let you know here when they are complete.

@a-vanderheiden any update on this? Please use the bot to generate pdf when you have successfully updated the `paper.md` file.
