### 2026-02-12 b061d25
added automatic creation of output_dir
