### 2026-02-12 8abaffa
write rasters to disk for later use
