### 2026-02-12 0ce94b1
added script to create plot for calculate ridge metrics doc
