### 2026-02-12 c729b9f
updated code to match plot script
