### 2026-02-12 28ea017
update noxfile with ridge metrics plots
