+++ noxfile.py
@@ -72,6 +72,7 @@ def docs(session: nox.Session) -> None:
     # Update Plots
     session.run("python", "scripts/plots/delineate_ridge_areas.py")
     session.run("python", "scripts/plots/create_vector_datasets.py")
+    session.run("python", "scripts/plots/calc_ridge_metrics.py")
 
     if serve:
         session.run("sphinx-autobuild", "--open-browser", *shared_args)
