### 2026-02-12 tamu-edu/scrollstats#34 @a-vanderheiden
@Atreyagaurav I just pushed the last changes to generate all of the example plots in the docs as part of the CI/CD to the [joss-review](https://github.com/tamu-edu/scrollstats/tree/joss-review) branch. These changes are now live in the [joss-review version of the docs](https://scrollstats.readthedocs.io/en/joss-review/index.html).
