### 2026-02-20 tamu-edu/scrollstats#35 @changliao1025
Installation error
I tried to install the package in a conda environment with python 3.12. and I got an error:

       File "<string>", line 67, in build_library_files
      TypeError: spawn() got an unexpected keyword argument 'dry_run'
      [end of output]

  ERROR: Failed building wheel for llvmlite
Failed to build llvmlite
error: failed-wheel-build-for-install

× Failed to build installable wheels for some pyproject.toml based projects
╰─> llvmlite
(scrollstats)
