### 2026-02-20 #9736 @changliao1025
Issue with installation: https://github.com/tamu-edu/scrollstats/issues/35
