### 2026-02-26 c38b4bc
added new required sections
