+++ paper/paper.md
@@ -82,6 +82,59 @@ has potential to also inform the predictions of future meander migration
 patterns and habitat suitability for riverine fauna and flora including riparian
 forests.
 
+# State of the field
+
+To the authors' best knowledge, there do not exist any other software packages
+purpose built to capture the variability in ridge morphometrics across scroll
+bar floodplains. The methodology and analysis of [@strick:2018] was influential
+in the creation of `ScrollStats`, but their analysis did not result in the
+creation of a software package, so contribution was not possible.
+
+`ScrollStats` is built upon the extensive scientific python ecosystem, and
+specifically relies upon popular geospatial libraries (shapely, geopandas, and
+rasterio) for spatial analysis and data manipulation. Users familiar with these
+python libraries should find working with and extending `ScrollStats` to be an
+intuitive experience.
+
+# Software design
+
+`ScrollStats` was built with interoperability and extensibility for the end-user
+in mind. For example, the `delineation` subpackage, which is responsible for
+delineating ridge areas from the input DEM, by default uses two classifier
+functions to delineate ridge areas within the DEM: profile curvature and
+residual topography. However, end users can extend ScrollStats by supplying
+their own list of classier functions that have the same callable signature as
+the default classifier functions:
+`classifier_func(ElevationArray2D, **kwargs) -> BinaryArray2D`. Likewise, the
+denoising process uses a default list of denoising functions that the user can
+extend with functions using the following callable signature
+`denoiser_func(BinaryArray2D, **kwargs) -> BinaryArray2D`
+
+Similarly, the `transecting` subpackage generates migration pathways using the
+vertical resultant calculations as described in [@hickin:1974]. However, if the
+user would prefer to use a different method of calculating migration
+trajectories from ridge to ridge, these alternative transects could be used as a
+drop-in replacement to calculate ridge metrics instead - so long as their
+vertices were coincident with the ridge lines they intersect.
+
+The DataExtractor classes in the `ridge_metrics` subpackage were designed to
+mirror the spatial scales at which they operate: Bend, Transect, and
+Intersection. This design communicates 1) which code is responsible for
+extracting information at the given scale and 2) what information is necessary
+as input to make these calculations. This enables future developers to
+easily identify where to focus efforts if they wish to extend the functionality
+of ScrollStats or troubleshoot unexpected results.
+
+# Research impact statement
+
+`ScrollStats` has had limited impact on research community at the time of
+publishing as it was not openly distributed beforehand. However, the core
+methodological framework of `ScrollStats` as well as initial findings from its
+use on meander bends from the Lower Brazos River, TX have been presented at the
+Association of American Geographers Annual Meeting in 2021 and 2023 during its
+development. Additionally, `ScrollStats` has been used in the completion of two
+Masters theses in the FLUD Lab at Texas A&M University.
+
 # Figures
 
 ![Ridgelines (dotted black) and channel centerline (solid grey) are manually digitized from interpretation of the DEM (Brazos River, Texas) and the binary ridge area raster (not pictured). `ScrollStats` then generates migration pathways (solid red) from equally spaced starting points along the centerline by "walking" up the floodplain from ridge to ridge (see Fig 3 from Hickin 1974 for transect generation procedure via calculation of vertical resultants). Ridge amplitude, width, and spacing are then calculated at each intersection (black dots) through analysis of the 1D signals generated by sampling both the DEM and binary ridge area raster along each transect. These calculations are shown for an example intersection (purple dot) of a ridge (solid blue) and migration pathway (thick solid red) in the following figures.\label{fig:figure1}](figs/digitized_dem.png)
@@ -91,6 +144,13 @@ forests.
[…]
