### 2026-02-26 28fdac9
added 'theses'  to codespell ignore list
