### 2026-02-26 #9736 @a-vanderheiden
Hi @Atreyagaurav, @changliao1025, @adithirgis I've added the newly required sections to the paper and pushed these changes to the branch [joss-review](https://github.com/tamu-edu/scrollstats/tree/joss-review). You can see the draft pdf at the link above.
