### 2026-02-26 tamu-edu/scrollstats#35 @a-vanderheiden
@changliao1025 I haven't tried installing scrollstats in a conda environment before, but I can check out this issue and see what I can find. Did you see this error with other versions of python or just 3.12?
