### 2026-02-26 tamu-edu/scrollstats#35 @a-vanderheiden
@changliao1025 I was able to reproduce the error when I tried to pip install scrollstats to a conda environment. I got the same error in a 3.10, 3.11, and 3.12 environment

It seems like the issue has to do with using pip to install numba (and therefore llvmlite) in a conda environment. According to the llvmlite docs, [pip installable wheels are only created for x86 architecture](https://llvmlite.pydata.org/en/latest/admin-guide/install.html#using-pip). So, if you try to pip install llvmlite on some other architecture (I get this same issue with my M1 mac), pip tries to build llvmlite for the wrong architecture and fail. They suggest all non x86 architecture users to either [use conda or build from source](https://llvmlite.pydata.org/en/latest/admin-guide/install.html#things-to-fix-it).

I was able to successfully install pip install scrollstats on non x86 architecture by first installing numba with conda:
```
conda create --name py312 python=3.12
conda activate py312

(py312) conda install -c conda-forge numba
```
Then pip installing scrollstats like normal
```
(py312) pip install scrollstats
```

Does this work for you?
