### 2026-02-26 tamu-edu/scrollstats#35 @changliao1025
Thank you for the update. Python 3.12 is rather stable version compared newer 3.13 and 3.14. So ibassumed something is wrong in the dependency chain. The package itself does not depends on the llvmlitebut somewhere else does.
I was testing using an iMac.

It would be great this can be fixed so major platforms can be supported as I don't see a reason why it cannot.
