### 2026-02-28 ccb9717
added conda install workaround
