### 2026-02-28 tamu-edu/scrollstats#35 @a-vanderheiden
@changliao1025 I agree, I'd like scrollstats to work easily with conda too. I'm looking into getting scrollstats installable directly from conda-forge. The [staging process](https://conda-forge.org/docs/maintainer/adding_pkgs/) may take me some time though to implement and get manual approval from the conda-forge review team.

For now though, I've added the above workaround to the [README on the`joss-reviews` branch](https://github.com/tamu-edu/scrollstats/tree/joss-review?tab=readme-ov-file#user-installation). What do you think?
