### 2026-03-05 #9736 @a-vanderheiden
Hi @adithirgis thanks for checking in. I believe I have addressed all of the issues raised by the reviewers so far.

The last thing I'm working on is getting scrollstats on conda-forge to address the issue with installing with conda ([issue 35](https://github.com/tamu-edu/scrollstats/issues/35)). In the mean time though, I have the workaround for installing with conda documented in the [README](https://github.com/tamu-edu/scrollstats/blob/joss-review/README.md#user-installation).
