### 2026-03-09 #9736 @changliao1025
- Since we already discussed the installation error, see previous comment. Please add the updated installation information to the documentation.
- I am able to run the simulation on my iMac after addressing the installation issue. That is great to see the software can run smoothly. However, as other comments suggested, there in no example code under the repo, I had to cope/paste from the document. Therefore, I suggest add the example code in the repo directly.
- When I looked at the input and output, I dont quite understand how the input data were obtained except the DEM. It appears some manual edits are involved. Can you provide more details on this. This is important because if a user apply this to a different region, then new input data must be prepared using a consistent approach.
