### 2026-03-12 #9736 @a-vanderheiden
Hi @changliao1025

> Since we already discussed the installation error, see previous comment. Please add the updated installation information to the documentation.

I've added the updated installation information to the [README on the joss-review branch](https://github.com/tamu-edu/scrollstats/tree/joss-review?tab=readme-ov-file#user-installation). My plan is to keep all of the commits from the review process on that branch then merge it with main to make a new release once the review is complete.

> I am able to run the simulation on my iMac after addressing the installation issue. That is great to see the software can run smoothly. However, as other comments suggested, there in no example code under the repo, I had to cope/paste from the document. Therefore, I suggest add the example code in the repo directly.

@Atreyagaurav mentioned this too in [issue #31](https://github.com/tamu-edu/scrollstats/issues/31). For this issue, I created the [scripts](https://github.com/tamu-edu/scrollstats/tree/joss-review/scripts) folder with example code that can be ran directly without needing to copy/paste the snippets from the docs.

> When I looked at the input and output, I don't quite understand how the input data were obtained except the DEM. It appears some manual edits are involved. Can you provide more details on this. This is important because if a user apply this to a different region, then new input data must be prepared using a consistent approach.

You're correct, some of the vector datasets need to be manually digitized from the DEM to be used in ScrollStats.  The [Create Vector Datasets](https://scrollstats.readthedocs.io/en/latest/CreateVectorDatasets.html) page has details on how each of the datasets for a bend are created/used. But I agree, the descriptions of each dataset could more clearly state which datasets are manually created. I'll update the first bulleted list on that page to more clearly state which data sets are manually vs automatically created and fix other typos that I saw there when I read through it.
