### 2026-03-12 23e5051
updated dataset descriptions with manual vs auto generation
