### 2026-03-12 301e939
removed references to python notebooks
