### 2026-03-12 afd175b
updated channel centerline paragraph
