### 2026-03-12 2489d23
updated note on vertex density
