+++ docs/CreateVectorDatasets.md
@@ -60,13 +60,14 @@ ridge metrics:
     along a bend according to the depositional record contained in the ridge and
     swale topography.
 
-**Note on point density for channel centerline and ridgelines**
+**Note on point density for channel centerline and ridge lines**
 
-> A high vertex density and high degree of "smoothness" are necessary for both
-> the channel centerline and ridge lines when creating the migration pathways.
-> It is recommended that you focus on capturing the overall ridge form, not
-> point density, when delineating. Then, the coarse polylines can be densified
-> and smoothed with the included [LineSmoother](scrollstats.LineSmoother) class.
+> The migration pathway algorithm requires a high vertex density and high degree
+> of "smoothness" for the input channel centerline and ridge line datasets. It
+> is recommended that you focus on capturing the overall form of these linear
+> features and not worry about point density when digitizing. The coarse polylines can be
+> densified and smoothed with the included
+> [LineSmoother](scrollstats.LineSmoother) class.
 
 ---
 
@@ -78,10 +79,7 @@ The details of use for each of these datasets are summarized below:
 
 **Required for Raster Processing**:
 
-- **Bend Boundary:** the bend boundary is a required input from the user and is
-  required for the raster processing. The bend boundary is the manual
-  delineation of the topographically raised platform that contains the ridge and
-  swale topography of a bend.
+- **Bend Boundary:** the bend boundary is a required to clip the ridge and swale topography from the DEM 
 
 **Required for Migration Pathway Creation**
 
