### 2026-03-12 49aa629
updated migration pathway requirements
