### 2026-03-12 ddd2860
updated ridge metric calculation requirements
