### 2026-03-12 2d06bfd
removed data management paragraphs
