### 2026-03-12 tamu-edu/scrollstats#36 @a-vanderheiden
Vector Dataset Creation
@changliao1025 I've updated the Create Vector Datasets page in the "joss-review" version of the documentation with more clear distinctions between manually vs automatically generated vector datasets for a bend. I've also fixed other errors/redundancies I found on that page.

You can access that specific version of the page at the link below. Please let me know what you think.

https://scrollstats.readthedocs.io/en/joss-review/CreateVectorDatasets.html

Related: [joss-reviews/issues/9736](https://github.com/openjournals/joss-reviews/issues/9736)
