### 2026-03-14 tamu-edu/scrollstats#35 @a-vanderheiden
@changliao1025 good news! I've opened a [PR](https://github.com/conda-forge/staged-recipes/pull/32406) to get scrollstats listed on conda-forge. Once it's accepted and merged by the conda-forge team, scrollstats should be installable directly with conda.
