### 2026-03-15 tamu-edu/scrollstats#38 @Atreyagaurav
[JOSS] Indication of integration into stablished research workflows or clear potential for reuse
Hi,

The paper.md has very little citations in terms of related studies. Can you expand more, maybe in paper.md, or in the documentation where you show similar research that could have taken advantage of this software, or future potential for reuse?

The need to make manual vector input files might deter people from using it easily.

I can understand the boundary, but other steps might be something that can be automated:
- center line: https://github.com/VeinsOfTheEarth/RivGraph
- ridge lines: peak detection algorithms for image/DEM
- migration pathways: perpendicular lines from ridges

Of course, you can decide based on the effort to automate, if more data is needed, and effort from the users needed.

I know automating them right not might not be possible, so I would like to see at least some thoughts and the roadmap for the future, along with plans for continuous development of the software.

Related: https://github.com/openjournals/joss-reviews/issues/9736
