### 2026-03-18 tamu-edu/scrollstats#38 @a-vanderheiden
Yes, I can add more cites of related studies that could potentially benefit from a tool like scrollstats. I'll begin working on that this week and let you know my progress here.

I agree, the manual creation of these vector datasets is not ideal. I have my thoughts for each that you brought up below

**Centerline:** Using a tool like RivGraph to generate a centerline is a good suggestion - especially for analysis of an entire river/reach. I'll add a link to it in the documentation. I would however still recommend that users with 1 to a few bends just roughly digitize the centerline from the DEM and smooth the output with the included [`LineSmoother`](https://github.com/tamu-edu/scrollstats/blob/b266e338a5f719ed0c944159d8cf068ffb393877/src/scrollstats/delineation/line_smoother.py#L9) class. This manual process may be easier than preparing a channel mask and installing a separate tool (depending on the circumstance).

**Ridge lines:** We actually spent a lot of time trying to automate the ridge line creation before releasing scrollstats, but couldn't find a solution that we were happy with and worked across many bends. We specifically tried using the binary ridge area raster with functions from skimage like [`medial_axis`](https://scikit-image.org/docs/stable/api/skimage.morphology.html#skimage.morphology.medial_axis) and [`skeletonize`](https://scikit-image.org/docs/stable/api/skimage.morphology.html#skimage.morphology.skeletonize), but the results would leave small line segments connected to the main ridge lines that would majorly divert the migration pathway algorithm. We experimented with a variety of methods to try and clean up the resulting lines, but none gave us the results we were looking for, so we decided to move forward with releasing the project without automated ridges. There are still many avenues to pursue to clean up the automated ridge lines (such as pruning/generalizing or looking into peak detection algorithms like you recommended). This is definitely something on the road map to work on in the future.

**Migration Pathways:** These are automatically generated by scrollstats. The user provides the centerline and ridge lines to the [`create_transects`](https://github.com/tamu-edu/scrollstats/blob/b266e338a5f719ed0c944159d8cf068ffb393877/src/scrollstats/transecting/transect.py#L1283) function along with some other parameters and scrollstats generates the transects using the vertical resultant method described in [Hickin 1974](https://doi.org/10.2475/ajs.274.4.414) (fig 3)

**Bend Boundary:** This interestingly may be possible to automate by calculating image [entropy](https://scikit-image.org/docs/0.25.x/auto_examples/filters/plot_entropy.html). The ridge area in the DEM has higher entropy than the surrounding floodplain which makes it possible to create a mask for this boundary. I admittedly have not spent much time investigating this, but it's something that I came across incidentally when working on the ridge delineation process.
