### 2026-03-18 tamu-edu/scrollstats#33 @a-vanderheiden
Sorry I missed you last comment @Atreyagaurav! Thanks for the follow up. I've created the pull request: [JOSS Review](https://github.com/tamu-edu/scrollstats/pull/39)
