### 2026-03-18 tamu-edu/scrollstats#38 @Atreyagaurav
This is great!

I know academia doesn't value negative results as much, but I think it would be really nice if you could summarize what you tried and what results you got (even if it was bad) in something like developer notes/ next steps somewhere (repo or documentation) so that it opens the path for future work into detecting the ridges. Since we have stream detection using DEM in hydrology (using flow direction and accumulation), I would like to think the ridge detection should work under similar principles (of course, I haven't actually tried). Seeing your results would be helpful.

As for the Bend Boundary, if you have to manually insert ridge lines, then you can probably use a convex hull for that polygon. Maybe buffer and/or smoothing as well if you want that.

But if you can automate the ridge, then both are automatic. But having the bend boundary manual, then only looking at ridges (automatically) within that boundary might help simplify the automatic ridge detection as well. Manually drawing the bend boundary might be preferable over manually drawing the ridges.
