### 2026-03-18 tamu-edu/scrollstats#41 @Atreyagaurav
Add architecture.md, or architecture diagram in contributing.md
Hi, can you add a description, ideally with diagrams, on the software architecture of scrollstats. Most of the documentation seems to be focused on the user side. Documentation focused on contributing developers --- that can give an overview of the inner mechanisms and which files/modules they should look for when they want to change specific things --- would be helpful.

Related: https://github.com/openjournals/joss-reviews/issues/9736
