### 2026-03-19 tamu-edu/scrollstats#38 @a-vanderheiden
I think sharing these negative results would be a good idea too. I'm thinking the repo's wiki may be a good platform for this kind of info. I can dig through my old notes and see what I can add there to document previous attempts at automated ridge line creation. I'm not familiar with stream detection methods from DEMs, so I'm not sure how similar that approach is to ridge detection here. Is there a resource you could point me to?

Using the convex hull to create the bend boundary polygon is interesting, but we found that you actually need the bend boundary first so that you can constrain which area of the binary raster gets automatically generated ridges. If you automatically create ridge lines for each ridge area identified in the [raster before clipping to the bend boundary](https://scrollstats.readthedocs.io/en/joss-review/_images/agreement_raster.png), you can see that you'd get MANY false ridges created.

Like you mentioned though, keeping the bend boundary manual (for now at least) will very likely simplify the automated ridge detection.
