### 2026-03-19 tamu-edu/scrollstats#40 @a-vanderheiden
This is interesting, thanks for the detailed notes. None of these issues seem straight forward to me though. The GitHub CI only goes up to python 3.12, but I'm not sure why 3.14 is giving an issue. I'll look into the following:
- if anything with python imports changed in 3.14
- if the lazy_load package supports/was changed in python 3.14
- if the nptyping package supports/was changed in python 3.14

As for test coverage for `calc_ridge_metrics`, I can make a test to cover that function. The tests do cover the different DataExtractor classes though which do all of the heavy lifting for ridge metric calculation - `calc_ridge_metrics` is just a wrapper around the `BendDataExtractor` class for convenience.
