### 2026-03-19 tamu-edu/scrollstats#38 @Atreyagaurav
> I'm not familiar with stream detection methods from DEMs, so I'm not sure how similar that approach is to ridge detection here. Is there a resource you could point me to?

Most search results just give me the tools for GIS softwares (e.g.: https://pro.arcgis.com/en/pro-app/latest/tool-reference/spatial-analyst/flow-accumulation.htm), but if you want to understand the theory, the steps are like this:

1. Fill depressions: to fix any "ponding" effect because the algorithm requires water to flow out completely, for your case breach depression might be better (it doesn't fill up the DEM),
2. Calculate Flow direction: this calculates which slope on the DEM is highest (where water is most likely to flow),
3. Calculate flow accumulation: this calculates the sum of all the cells that will flow to a particular cell,
4. streamline detection: take a threshold value, all cells with greater accumulation than this is a stream.

In case of ridges, all cells with flow accumulation of 0 (no other cell should flow to that cell) should be ridge cells. Or you can invert the DEM, then the detected streamlines are ridges in the original DEM.

In the figure below (from: https://docs.qgis.org/3.44/en/docs/training_manual/processing/hydro.html) shows flow accumulation results (log transform), the bright cells are those with high accumulation (streams), while the blue cells are those with low accumulation (ridges)

![accumulation](https://docs.qgis.org/3.44/en/_images/catchmentlayerlog.png)

Whitebox tools is an open source software so you can refer to it for how to implement different flow based algorithms:
https://www.whiteboxgeo.com/manual/wbt_book/available_tools/hydrological_analysis.html

> but we found that you actually need the bend boundary first so that you can constrain which area of the binary raster gets automatically generated ridges

I meant it in a way that you don't need both polygon and ridge lines manual.

Something like:

| Boundary | Ridge Lines | Case                                                      |
| ---------- | ------------- | ----------------------------------------------------------- |
| Manual   | Manual      | redundant                                                 |
| Auto     | Manual      | Worst case (boundary = convex hull of manual ridge lines) |
| Manual   | Automatic   | Expected use (ridge = clip of ridge detection inside boundary)    |
| Auto     | Auto        | Ideal case                                                |

But yeah, even if there are methods, it will take some time to test them out, and it might not work in this case. Documenting it is generally a good idea.

Edit: The details on stream extraction is from university courses, so maybe searching for "GIS in Hydrology" courses, or similar will give you a good resources.
