### 2026-03-21 1bf7269
added ARCHITECTURE.md with project tree
