### 2026-03-21 0088b17
added Data Stores section to ARCHITECTURE.md
