### 2026-03-22 eb548e3
added high-level diagram
