### 2026-03-24 5bbf203
added items to ARCHITECTURE.md
