### 2026-03-24 tamu-edu/scrollstats#41 @a-vanderheiden
@Atreyagaurav I've created an [ARCHITECTURE.md](https://github.com/tamu-edu/scrollstats/blob/arch/ARCHITECTURE.md) file using [https://architecture.md](https://architecture.md/) as a guide.

In there I've added a mermaid diagram showing the general flow from DEM -> ridge metrics. I still need to add descriptions of the core components in scrollstats (section 3), but this is nearly finished.
