### 2026-03-24 tamu-edu/scrollstats#38 @a-vanderheiden
@Atreyagaurav thanks for the explanation! I think this flow accumulation technique (where flow accumulation is 0) is worth investigating as a potential method for auto ridge generation. I can add this to the list of future methods to try. I've created a public [project](https://github.com/users/a-vanderheiden/projects/5) where I'll be tracking progress on scrollstats development. This will go there soon.
