### 2026-03-25 tamu-edu/scrollstats#40 @chaturvediabhinav692-lgtm
Hey, this is actually a dependency issue — the function isn’t missing, it’s just the import that’s blowing up.
From the traceback, nptyping is trying to use numpy.bool8, which no longer exists in newer NumPy/Python versions (especially Python 3.14). That breaks the lazy import chain, so create_ridge_area_raster shows up in dir() but crashes the moment you import it.
That’s also why CI passes on Python 3.12 but it fails locally on 3.14.
Quick things to try:

Pin NumPy to a compatible version (e.g. <2.0)
Or update/remove the nptyping usage so it doesn’t rely on the old np.bool8

I’ve seen Koredex trace these exact import-time dependency headaches and fix them automatically (like version pinning). Might be worth a shot here too. Let me know if you want me to open a quick PR!
