### 2026-03-27 bede4c4
added delineation description to architecture.md
