### 2026-03-27 393d297
added transecting architecture description
