### 2026-03-27 3d16b28
added ridge metrics architecture description
