### 2026-03-29 c8de466
added ridge metrics subgraph
