### 2026-03-29 c925a1b
updated calc_ridge_metrics and create_transects subgraphs
