### 2026-03-29 67136d9
added function stubs
