### 2026-03-29 tamu-edu/scrollstats#41 @a-vanderheiden
@Atreyagaurav [ARCHITECTURE.md](https://github.com/tamu-edu/scrollstats/blob/joss-review/ARCHITECTURE.md) is now complete and merged into the joss-review branch.
