### 2026-03-29 497227e
removed nptyping from array_types
