### 2026-03-29 ed2cc90
removed nptyping as dependency
