+++ pyproject.toml
@@ -35,7 +35,6 @@ dependencies = [
   "rasterio",
   "numba",
   "tqdm",
-  "nptyping",
   "lazy-loader"
   ]
 
