### 2026-03-29 tamu-edu/scrollstats#40 @a-vanderheiden
@chaturvediabhinav692-lgtm thanks for the explanation and your interest in ScrollStats! I've gone ahead and removed the nptyping module from ScrollStats altogether and replaced the array types that used it with the generic `numpy.ndarray`

@Atreyagaurav I've also added python 3.13 and 3.14 to the CI to catch any other issues that may arise with the latest versions of python.
