### 2026-03-30 fa985f0
added test for calc_ridge_metrics
