### 2026-03-30 tamu-edu/scrollstats#40 @a-vanderheiden
@Atreyagaurav I've [added the test](https://github.com/tamu-edu/scrollstats/pull/39/changes/fa985f0adba530e1b1e427651fc50f8db72f8a1b) for `calc_ridge_metrics()`
