### 2026-04-02 tamu-edu/scrollstats#38 @a-vanderheiden
@Atreyagaurav I've written up the negative results from past attempts at automatic ridge line generation and added it to the [Developer Notes](https://github.com/tamu-edu/scrollstats/wiki/Developer-Notes) page on the wiki!
