### 2026-04-02 #9736 @a-vanderheiden
Hi @adithirgis apologies for the delay, the review process is still going well. I recently wrapped up work on issues [40](https://github.com/tamu-edu/scrollstats/issues/40) and [41](https://github.com/tamu-edu/scrollstats/issues/41), and now I'm working on issue [38](https://github.com/tamu-edu/scrollstats/issues/38). I just finished writing up the requested results from previous attempts to generate ridge lines manually and my next task is to add more citations to paper.md for potential use of ScrollStats.
