### 2026-04-10 32dbf91
added references to paper.md
