+++ paper/paper.md
@@ -53,9 +53,11 @@ bars preserved in the ridge and swale topography commonly found in the
 floodplains of meandering rivers adjacent to the river channel. This
 quantification will allow researchers to investigate the relationships between
 ridge morphology and the environmental factors affecting its formation, such as
-the hydrology at the time of deposition, spatial variations in the river width,
-the channel curvature, the position along the meander bend, and the floodplain
-vegetation coverage and composition.
+the hydrology at the time of deposition[@nanson:1980], spatial variations in the
+river width [@van_de_lageweg:2014; @strick:2018], the channel
+curvature[@strick:2018], the position along the meander bend[@nanson:1980],
+anthropogenic modifications to the channel[@nagy:2020], and the floodplain
+vegetation coverage and composition [@hickin:1975; @nanson:1977; @zen:2017].
 
 `ScrollStats` generates a series of migration pathways (an adaptation of the
 "erosion pathlines" from @hickin:1974) that trace the paths of migration across
@@ -66,6 +68,10 @@ one-dimensional (1-D) signals of ridge elevation and ridge presence
 (\autoref{fig:figure2}). Then, from each 1-D signal, the ridge's amplitude,
 width, and spacing (distance from the previous ridge) can be calculated at every
 intersection of the migration pathway and a ridge (\autoref{fig:figure3}).
+Calculating these metrics for multiple migration pathways may improve upon
+results from previous studies that calculated similar metrics for a single
+transect per bend [@strick:2018; @nagy:2020] by capturing the spatial
+variability in these metrics across the bend.
 
 The intersections of migration pathways and ridge lines form a migrationally
 relevant grid, which allows for the measurements at each intersection to be
@@ -76,11 +82,12 @@ ridge) and along the channel (from migration pathway to migration pathway) and
 examine the associations between these changes in ridge morphology and the
 environmental factors affecting their formation. This allows for researchers to
 leverage the morphological information stored in the floodplains of meandering
-rivers to deduce past events such as changes in flow regimes, river planform and
-bend dynamics, sediment flux, and carbon storage and release. Such information
-has potential to also inform the predictions of future meander migration
-patterns and habitat suitability for riverine fauna and flora including riparian
-forests.
+rivers to deduce past events such as changes in flow regimes [@bridge:1995;
+@van_de_lageweg:2014], river planform and bend dynamics [@nanson:1980;
+@lapointe:1986], sediment flux[@nanson:1980], and carbon storage and release.
+Such information has potential to also inform the predictions of future meander
+migration patterns and habitat suitability for riverine fauna and flora
+including riparian forests [@nanson:1977; @marks:2014].
 
 # State of the field
 
