### 2026-04-13 tamu-edu/scrollstats#38 @Atreyagaurav
The image shown here looks like you could remove the smaller segments by using an algorithm to extract the longest continuous line. Then run a smoothing algorithm to fix the sudden changes.  The smoothing can also be applied with a weighted 'aligning' factor that can make sure the line moves similarly to the lines that are its immediate neighbors. That should fix the issue you had with the migration paths going haywire due to the detected ridges. You probably have to test different weights for smoothing and aligning to get a good result without impacting the actual ground truth too much, but it should be close to manually drawing (and better in terms of efforts).

![image](https://private-user-images.githubusercontent.com/70538538/572879355-2be50444-9f1e-4c55-b4e3-1dfa332f5bca.png?jwt=eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3NzYwOTk2MzMsIm5iZiI6MTc3NjA5OTMzMywicGF0aCI6Ii83MDUzODUzOC81NzI4NzkzNTUtMmJlNTA0NDQtOWYxZS00YzU1LWI0ZTMtMWRmYTMzMmY1YmNhLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNjA0MTMlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjYwNDEzVDE2NTUzM1omWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTM0OGI0ZDQ0ZjExMTg1NzE4OGY4ZDZiNjNmYWYwOWNjMDdiMDFlMzBjYjcxODI0OGUyNGYxZGQ4MGM2ZWE4NmUmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JnJlc3BvbnNlLWNvbnRlbnQtdHlwZT1pbWFnZSUyRnBuZyJ9.CY8E_El2_LqzxLKzhaprrrVDSj_gdHpsnat6GfxDQ5o)
