### 2026-04-15 e16c57a
add citation entries to paper.bib
