+++ paper/paper.bib
@@ -1,3 +1,13 @@
+@article{bridge:1995,
+	title = {Ground-penetrating radar and coring used to study the large-scale structure of point-bar deposits in 3 dimensions},
+	volume = {42},
+	doi = {10.1111/j.1365-3091.1995.tb00413.x},
+	pages = {839--852},
+	number = {6},
+	journal = {Sedimentology},
+	author = {Bridge, J. S. and Alexander, J. and Collier, R. E. L. and Gawthorpe, R. L. and Jarvis, J.},
+	date = {1995},
+}
 
 @article{hickin:1974,
 	title = {The development of meanders in natural river-channels},
@@ -11,6 +21,39 @@
 	langid = {english},
 }
 
+@article{hickin:1975,
+	title = {The Character of Channel Migration on the {Beatton} {River}, {Northeast} {British} {Columbia}, {Canada}},
+	volume = {86},
+	doi = {10.1130/0016-7606(1975)86<487:TCOCMO>2.0.CO;2},
+	pages = {487--494},
+	number = {4},
+	journaltitle = {GSA Bulletin},
+	author = {Hickin, Edward J. and Nanson, Gerald C.},
+	date = {1975},
+}
+
+@article{lapointe:1986,
+	title = {Migration patterns of an asymmetric meandering river - {The} {Rouge} {River}, {Quebec}},
+	volume = {22},
+	doi = {10.1029/WR022i005p00731},
+	pages = {731--743},
+	number = {5},
+	journal = {Water Resources Research},
+	author = {Lapointe, M. F. and Carson, M. A.},
+	date = {1986},
+}
+
+@article{marks:2014,
+	title = {Quantifying flooding regime in floodplain forests to guide river restoration},
+	volume = {2},
+	doi = {10.12952/journal.elementa.000031},
+	pages = {000031},
+	number = {0},
+	journaltitle = {Elementa: Science of the Anthropocene},
+	author = {Marks, Christian O. and Nislow, Keith H. and Magilligan, Francis J.},
+	date = {2014},
+}
+
 @article{nagy:2020,
 	title = {Point-bar Development Under Human Impact: Case Study on the {Lower Tisza River, Hungary}},
 	volume = {24},
@@ -22,6 +65,28 @@
 	date = {2020},
 }
 
+@article{nanson:1977,
+	title = {Forest Succession and Sedimentation on a Meandering-River Floodplain, {Northeast} {British} {Columbia}, {Canada}},
+	volume = {4},
+	doi = {10.2307/3038059},
+	pages = {229--251},
+	number = {3},
+	journaltitle = {Journal of Biogeography},
+	author = {Nanson, Gerald C. and Beach, Harry F.},
+	date = {1977},
+}
+
+@article{nanson:1980,
+	title = {Point bar and floodplain formation of the meandering {Beatton} {River}, northeastern {British} {Columbia}, {Canada}},
+	volume = {27},
+	doi = {10.1111/j.1365-3091.1980.tb01155.x},
+	pages = {3--29},
+	number = {1},
+	journaltitle = {Sedimentology},
+	author = {Nanson, Gerald C.},
+	date = {1980},
+}
+
 @article{strick:2018,
 	title = {Morphology and spacing of river meander scrolls},
 	volume = {310},
@@ -30,6 +95,17 @@
 	journaltitle = {Geomorphology},
 	author = {Strick, Robert J. P. and Ashworth, Philip J. and Awcock, Graeme and Lewin, John},
 	date = {2018},
+}
+
+@article{van_de_lageweg:2014,
+	title = {Bank pull or bar push: What drives scroll-bar formation in meandering rivers?},
+	volume = {42},
[…]
