+++ paper/paper.bib
@@ -104,7 +104,7 @@
 	pages = {319--322},
 	number = {4},
 	journaltitle = {Geology},
-	author = {van de Lageweg, Wietse I. and van Dijk, Wout M. and Baar, Anne W. and Rutten, Jantien and Kleinhans, Maarten G.},
+	author = {{van de Lageweg}, Wietse I. and {van Dijk}, Wout M. and Baar, Anne W. and Rutten, Jantien and Kleinhans, Maarten G.},
 	date = {2014},
 }
 
