### 2026-04-15 #9736 @a-vanderheiden
Hi @adithirgis, the review process is still going well. It took me longer than expected, but I just pushed changes to paper.md for issue [38](https://github.com/tamu-edu/scrollstats/issues/38).
