### 2026-04-15 tamu-edu/scrollstats#38 @a-vanderheiden
@Atreyagaurav you're correct in that the spurs will have to be removed from the skeletonized outputs. When we tried spur removal in the past, it did improve migration pathway depth - but not nearly to the degree that we were looking for. Perhaps that extra alignment step you mentioned could help too. I'm not familiar with this technique though.

The smoothing algorithm currently in use by scrollstats for manually created ridges is a simple mean filter on the xy coords of a single line. It isn't aware of any nearby lines. Do you know of an example implementation of this neighbor alignment method you could point me to?
