### 2026-04-15 tamu-edu/scrollstats#38 @a-vanderheiden
@Atreyagaurav I also just pushed edits to paper.md with additional citations showing how similar research could have used scrollstats and demonstrating future potential for reuse. You can see the changes in PR #44 and the [new draft PDF](https://github.com/openjournals/joss-papers/blob/joss.09736/joss.09736/10.21105.joss.09736.pdf) I generated with editorialbot.
