### 2026-04-18 tamu-edu/scrollstats#38 @Atreyagaurav
Hi,

Sorry, I don't have an example to show how smoothing with neighbors work. But it is basically the same step with smoothing, but you take a "slope" or "direction" of all the lines in the area, or neighbors (just left and right) and use that to adjust the direction of the current line during smoothing. You can assign weights to see how much aligned you want. One extreme is no change (0 weight) while the other extreme results in lines that are equal distance from each other throughout the running length.

We can probably continue this discussion after the current review process as well. Let me see if I can send something your way later.
