### 2026-04-18 tamu-edu/scrollstats#45 @Atreyagaurav
[JOSS] Citation format
Please change this citation to a narrative citation format (with parenthesis on the year and not on author). You have a correct example of citation in text in line 30.

> https://pandoc.org/MANUAL.html#extension-citations

Same problem in Line: 72

Related: https://github.com/openjournals/joss-reviews/issues/9736
