### 2026-04-18 #9736 @Atreyagaurav
Hi @adithirgis,

I have completed my review. The above issue is the last one; once that is closed, the submitting author would have satisfied all my requests.

I have some concerns about the open development history, and the future reusability of the software. But as you know, those were not hard requirements when the paper was submitted. And I hope the authors will continue the development and increase the usefulness of the package in the future.

Edit: the author has fixed the issues with citation format, so all good from me.
