### 2026-04-18 2fa3480
changed citations to narrative format
