### 2026-04-18 tamu-edu/scrollstats#45 @a-vanderheiden
@Atreyagaurav as you have [already seen](https://github.com/openjournals/joss-reviews/issues/9736#issuecomment-4272539390), I've updated the mentioned citations to use narrative format in the [latest draft](https://github.com/openjournals/joss-papers/blob/joss.09736/joss.09736/10.21105.joss.09736.pdf).

Thanks for pointing that out!
