### 2026-04-18 tamu-edu/scrollstats#38 @a-vanderheiden
@Atreyagaurav That's alright, I'm sure I can find some examples online too. Conceptually though, I think I understand the process you're describing.

I'm looking forward to adding this functionality after the review though! Manually digitizing ridge lines was by far the most laborious part of prepping the data.
