### 2026-04-18 tamu-edu/scrollstats#38 @Atreyagaurav
Yeah, and trying to automate such a thing is frustrating. Trying to cover all edge cases is hard. And when your automation takes way more time than it would have taken to do it manually, it becomes hard to keep going.

While manually doing it can help you complete the current task, making an automated way to do that can contribute to the science and help many people that come after you. But again, you have to choose one or the other based on your project constraints.

Good luck with the rest of the review. Again, feel free to ping me if you need any help implementing this later. As you have probably seen, I have experience with digital elevation models in hydrology, and river network detection and analysis, so this seems like a fun problem to solve (and definitely solvable; I don't know how much effort will be required though :D ).
