### 2026-04-18 555597d
updated conda install instructions; resolves #35
