### 2026-04-18 tamu-edu/scrollstats#35 @a-vanderheiden
@changliao1025 apologies for not following up with you on this, but scrollstats is now installable with conda directly using
```
conda install scrollstats
```

There is a catch however. Directly installing with conda only works for python <=3.12 at the moment. When I try to install with python 3.13, I get the following error stating that there is a version incompatibility between `numpy` and `nptyping` at python 3.13+.

```
(base) avan@AndrewsMBP dev % conda create -n scroll313 python=3.13
(base) avan@AndrewsMBP dev % conda activate scroll313
(scroll313) avan@AndrewsMBP dev % conda install scrollstats
Channels:
 - conda-forge
 - defaults
Platform: osx-64
Solving environment: failed

LibMambaUnsatisfiableError: Encountered problems while solving:
  - package scrollstats-0.1.3-pyhcf101f3_0 requires nptyping, but none of the providers can be installed

Could not solve for environment specs
The following packages are incompatible
├─ pin on python 3.13.* =* * is installable and it requires
│  └─ python =3.13 *, which can be installed;
└─ scrollstats =* * is not installable because it requires
   └─ nptyping =* *, which requires
      └─ numpy [<2 *|<2.0 *] but there are no viable options
         ├─ numpy [1.21.2|1.21.3|...|1.26.4] would require
         │  └─ python >=3.10,<3.11.0a0 *, which conflicts with any installable versions previously reported;
         ├─ numpy [1.22.3|1.23.4|...|1.26.4] would require
         │  └─ python >=3.11,<3.12.0a0 *, which conflicts with any installable versions previously reported;
         ├─ numpy [1.11.3|1.13.3|...|1.24.4] would require
         │  └─ python >=3.8,<3.9.0a0 *, which conflicts with any installable versions previously reported;
         ├─ numpy [1.16.6|1.19.2|...|1.26.4] would require
         │  └─ python >=3.9,<3.10.0a0 *, which conflicts with any installable versions previously reported;
         ├─ numpy [1.26.0|1.26.2|1.26.3|1.26.4] would require
         │  └─ python [>=3.12,<3.13.0a0 *|>=3.12.0rc3,<3.13.0a0 *], which conflicts with any installable versions previously reported;
         ├─ numpy [1.10.4|1.11.0|...|1.9.3] would require
         │  └─ python [=2.7 *|>=2.7,<2.8.0a0 *], which conflicts with any installable versions previously reported;
         ├─ numpy [1.10.4|1.11.0|...|1.9.3] would require
         │  └─ python =3.4 *, which conflicts with any installable versions previously reported;
         ├─ numpy [1.10.4|1.11.0|...|1.9.3] would require
         │  └─ python [=3.5 *|>=3.5,<3.6.0a0 *], which conflicts with any installable versions previously reported;
         ├─ numpy [1.10.4|1.11.3|...|1.9.3] would require
         │  └─ python =3.6 *, which conflicts with any installable versions previously reported;
         ├─ numpy [1.11.3|1.12.1|...|1.9.3] would require
         │  └─ python >=3.6,<3.7.0a0 *, which conflicts with any installable versions previously reported;
         └─ numpy [1.11.3|1.14.5|...|1.9.3] would require
            └─ python >=3.7,<3.8.0a0 *, which conflicts with any installable versions previously reported.

Pins seem to be involved in the conflict. Currently pinned specs:
 - python=3.13
```

This however will not be an issue once I merge PR #39 and bump the release version of scrollstats to v0.1.4 because I've since removed `nptyping` as a dependency (see issue #40).

Once the new version is released to PyPI, a PR will automatically open on the conda-forge [scrollstats-feedstock](https://github.com/conda-forge/scrollstats-feedstock/tree/main) repo to bump the version there too. I'll then remove [`nptyping`](https://github.com/conda-forge/scrollstats-feedstock/blob/5a4295f5db856fea81e09a865dd39118e6485483/recipe/recipe.yaml#L33) from the recipe there and merge the PR, removing the error above.

I've updated the install instructions in the [README](https://github.com/tamu-edu/scrollstats/blob/joss-review/README.md#user-installation) to install directly with conda and that scrollstats v0.1.3 requires python <=3.12.
