### 2026-04-18 tamu-edu/scrollstats#36 @a-vanderheiden
@changliao1025 have you had a chance to check out this page? I believe my edits there have addressed your [concerns](https://github.com/openjournals/joss-reviews/issues/9736#issuecomment-4025675865) mentioned in the main review thread.
