### 2026-04-18 tamu-edu/scrollstats#38 @a-vanderheiden
Totally, I think you summed up the problem there pretty well. All the time spent chasing down all of the edge cases was what drove us to go the manual route in the first place.

And thanks for the offer! I'll take you up on that. I have a few other ideas that I would like to try too once the review is complete. I just haven't had the time to work on them while I've been working on everything else with this package for the submission.

Thanks again for all your help and guidance! I've learned a lot through this process and I really appreciate the time you spent on this review while prepping for your defense.
