### 2026-04-22 c5dcbd2
Update README with a bend DEM visualization
