### 2026-04-28 #9736 @adithirgis
> Hi [@adithirgis](https://github.com/adithirgis),
>
> I have completed my review. The above issue is the last one; once that is closed, the submitting author would have satisfied all my requests.
>
> I have some concerns about the open development history, and the future reusability of the software. But as you know, those were not hard requirements when the paper was submitted. And I hope the authors will continue the development and increase the usefulness of the package in the future.
>
> Edit: the author has fixed the issues with citation format, so all good from me.

Hello @Atreyagaurav - thank you for the thorough review! Regarding your concern about open development history, and the future reusability of the software - would you please share in detail regarding the same.

@a-vanderheiden do you have any thoughts on this -

> I have some concerns about the open development history, and the future reusability of the software. But as you know, those were not hard requirements when the paper was submitted. And I hope the authors will continue the development and increase the usefulness of the package in the future.

Thank you both!
