### 2026-04-28 #9736 @changliao1025
@a-vanderheiden has addressed most of my concerns already, especially the installation section.
After this round, I think the software is in a much better shape for other users to test and apply the software to other domains.

My last recommendation is that can you put the installation instruction in the documentation as well other than the readme file?

Thank you.
