### 2026-04-28 tamu-edu/scrollstats#35 @changliao1025
Dependency is always an issue, I am glad you found a better solution to address this.
