### 2026-04-29 #9736 @Atreyagaurav
Hi @adithirgis ,

Here are more details on my two general concerns.

# Reusability

My concern with future reusability is based on the fact that many inputs required for this software need manual work. They neither have any prior methods to generate the input data nor any publicly available dataset to download the inputs from. That means, someone must invest their time and effort to generate specific input data before being able to use this software.

I have suggested to the author that they automate the input data generation (refer: https://github.com/tamu-edu/scrollstats/issues/38). The author has added documentation related to the methods they have attempted, and explained how they were unable to automate it at the moment.

I expect future developments to address that in order to increase the reusability of this project. As we expect all JOSS projects to continue development, I think that is not a problem.

#  Open development
My concern for the Open development is based on this: https://joss.readthedocs.io/en/latest/review_criteria.html#open-development

> The software should have been developed openly from early stages. For projects with recently created public repositories, > there should be at least six months of public development history with:
> Evidence of releases or version tags
> Public issues and/or pull requests
> Ideally, external engagement from users or contributors outside the core team

>  Good: Repository public from inception with documented releases and community interaction
>  OK: Repository made public 6+ months ago with clear evidence of ongoing development
>  Not acceptable: Repository made public immediately before submission with limited public development history

This repository does not have a long history of open development, and based on this: https://github.com/tamu-edu/scrollstats/issues/18, I believe the repo was made public just before submission.

But I learned that the guidelines were updated after the submission of this paper, so I have no issue with that.

Previous guidelines: https://joss.readthedocs.io/en/pre-2026-submission/review_criteria.html#substantial-scholarly-effort
