### 2026-04-29 #9736 @changliao1025
I also raised the similar issue on the input data: https://github.com/tamu-edu/scrollstats/issues/36
More future development will be required for this software to become truly useful. But at least other users can understand what steps they can follow to manually create these input data.
