### 2026-04-30 #9736 @a-vanderheiden
Hi @adithirgis,

Both @Atreyagaurav and @changliao1025 are correct; the repo was developed privately then made public for submission to JOSS. As @Atreyagaurav mentioned though, the review criteria of 6+ months of open development was not required when I originally submitted ScrollStats. Additionally, since my initial submission in early 2025, ScrollStats has been developed publicly with versioned and tagged releases, based on input from the reviewers and editors here on this thread and the PRE-REVIEW thread, and in accordance with scientific python development best practices from https://learn.scientific-python.org/development/.

Both reviewers are also correct about the manual creation of certain input datasets required for ScrollStats. As recommended by @Atreyagaurav , the previous attempts made by my co-authors and I to automate the creation of these vector datasets from the DEM have been documented and shared publicly on the repo wiki so that other developers may see and build upon that work (please see https://github.com/tamu-edu/scrollstats/wiki/Developer-Notes ). Future development of ScrollStats will address the automated creation of these input datasets from the DEM (specifically the ridge lines) and the input provided by @Atreyagaurav on [issue 38](https://github.com/tamu-edu/scrollstats/issues/38) will likely be very helpful in that pursuit.

Until we have reliable automated methods to create these input datasets, I have made clear distinctions in the [documentation](https://scrollstats.readthedocs.io/en/joss-review/CreateVectorDatasets.html) between which of the vector datasets need to be created manually (as recommended by @changliao1025) and provided instructions on how to do so. In addition to the written instructions, the [example dataset](https://github.com/tamu-edu/scrollstats/tree/main/example_data/input) packaged with ScrollStats provides users with a complete and concrete example of these datasets.
