### 2026-05-01 #9736 @a-vanderheiden
>My last recommendation is that can you put the installation instruction in the documentation as well other than the readme file?

@changliao1025 Yes, I can add the install instructions to the docs too. Sorry I didn't get back to this comment yesterday.

Edit: the install instructions have been added to the [joss-review](https://scrollstats.readthedocs.io/en/joss-review/) version of the docs
