### 2026-05-01 55572a8
added install instructions to docs (#35)
