### 2026-05-08 2c71556
updated intro text
