### 2026-05-19 #9736 @adithirgis
Hi @a-vanderheiden the affiliations on Zenodo should be same as in the JOSS paper draft. Please update Zenodo archive. Thanks!
