### 2026-05-29 #9736 @kthyng
@a-vanderheiden
Paper comments:
* “mitigate future erosion” should probably be “mitigates future erosion” for subject–verb agreement in the Summary section.
* “end users can extend ScrollStats by supplying their own list of classier functions” appears to contain a typo; “classier” should likely be “classifier”.
* “Aggregate values represent the medain value” seems to have a typo
